% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\name{aggregate_kernel}
\alias{aggregate_kernel}
\title{Workhorse function for multi-stage aggregation}
\usage{
aggregate_kernel(data, group_by, all_cols, target_cols, fn, agg_unique,
  na = "na.omit")
}
\arguments{
\item{data}{object of class 'data.frame'.}

\item{group_by}{column names to use for grouping}

\item{all_cols}{all columns of data.frame to be used}

\item{target_cols}{column names to be aggregated}

\item{fn}{function to use for aggregation}

\item{agg_unique}{should the resulting data be unique?}

\item{na}{value passed to aggregate(..., na.action = na)}
}
\value{
\code{object} object of class 'DataSet'.
}
\description{
Workhorse function for multi-stage aggregation
}

