% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\name{aggregateData}
\alias{aggregateData}
\title{Aggregates new data with new columns to dataset}
\usage{
aggregateData(object, new_data)
}
\arguments{
\item{object}{object of class 'DataSet'.}

\item{new_data}{data.frame with new data}
}
\value{
\code{object} object of class 'DataSet'
}
\description{
Aggregates new data with new columns to dataset
}

