% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\name{aggregateGroup}
\alias{aggregateGroup}
\title{Runs aggregate on data stored in Dataset}
\usage{
aggregateGroup(object, target_cols, group_by, fn, agg_unique = FALSE)
}
\arguments{
\item{object}{object of class 'DataSet'.}

\item{target_cols}{column names to be aggregated}

\item{group_by}{column names to use for grouping}

\item{fn}{function to use for aggregation}

\item{agg_unique}{should the resulting data be unique?}
}
\value{
\code{agg} object of class 'data.frame'.
}
\description{
Runs aggregate on data stored in Dataset
}

