% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\name{aggregateGroupCompound}
\alias{aggregateGroupCompound}
\title{Runs multiple aggregate on data stored in Dataset within groups
defined by group_by_lst}
\usage{
aggregateGroupCompound(object, target_cols, group_by_lst, fn_lst,
  agg_unique = FALSE)
}
\arguments{
\item{object}{object of class 'DataSet'.}

\item{target_cols}{column names to be aggregated}

\item{group_by_lst}{list of  column names to use for sequential grouping}

\item{fn_lst}{function to use for aggregation on each step}

\item{agg_unique}{should the resulting data be unique?}
}
\value{
\code{agg} object of class 'data.frame'.
}
\description{
Runs multiple aggregate on data stored in Dataset within groups
defined by group_by_lst
}

