% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_dataplex.r
\name{analysis_module_load_multiple}
\alias{analysis_module_load_multiple}
\title{Load and bind analysis output from multiple stores}
\usage{
analysis_module_load_multiple(traders, dates, name_or_builder, key_fn)
}
\arguments{
\item{traders}{integer, trader ID's.}

\item{dates}{Date vector of dates}

\item{name_or_builder}{function or function name of builder used
to generate analysis module}

\item{key_fn}{function, key generator}
}
\value{
\code{posn_comp_obj} object of class 'PositionComposite'.
}
\description{
Queries analysis store objects for requested dates.
Retrieves or updates analysis and binds results together.
}

