% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_dataplex.r
\name{analysis_module_request}
\alias{analysis_module_request}
\title{query for analysis module for specific set of parameters}
\usage{
analysis_module_request(key_function, name_or_builder, force = FALSE)
}
\arguments{
\item{key_function}{"function" generating key}

\item{name_or_builder}{"character" name of the objectstore or "function" builder of the objectstore}

\item{force}{"logical" should Analysis be create if not existing in store}
}
\value{
\code{an_store} object of class "AnalysisObjectStore"
}
\description{
query for analysis module for specific set of parameters
}

