% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_analysis_objectstore.r
\name{analysis_objectstore_factory}
\alias{analysis_objectstore_factory}
\title{get analysis objectstore for given name}
\usage{
analysis_objectstore_factory(name)
}
\arguments{
\item{name}{"character" hashed name of the objectstore}
}
\value{
\code{anstr} object of class "AnalysisObjectStore"
}
\description{
get analysis objectstore for given name
}

