% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_dataplex.r
\name{analysis_store_request}
\alias{analysis_store_request}
\title{get analysis store for given key function}
\usage{
analysis_store_request(key_function)
}
\arguments{
\item{key_function}{"function" generating key}
}
\value{
\code{an_store} object of class "AnalysisObjectStore"
}
\description{
get analysis store for given key function
}

