% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_analysis_objectstore.r
\docType{class}
\name{AnalysisQuery-class}
\alias{AnalysisQuery-class}
\title{An S4 class handling queries to AnalysisObjectstore.}
\description{
An S4 class handling queries to AnalysisObjectstore.
}

