% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\name{appendData}
\alias{appendData}
\title{Append new data to dataset}
\usage{
appendData(object, new_data)
}
\arguments{
\item{object}{object of class 'DataSet'.}

\item{new_data}{data.frame with new data}
}
\value{
\code{object} object of class 'DataSet'
}
\description{
Append new data to dataset
}
\examples{
appendData(object,new_data)
}

