% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\name{attemptUpdate}
\alias{attemptUpdate}
\title{Safely attempts to update dataset with new data}
\usage{
attemptUpdate(object, data, new_data)
}
\arguments{
\item{object}{object of class 'DataSet'.}

\item{data}{data.frame with new dataset data to be set combined with existing data}

\item{new_data}{data.frame with original new data}
}
\value{
\code{object} object of class 'DataSet'
}
\description{
Safely attempts to update dataset with new data
}

