% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urldatareader.r
\name{autoSchema}
\alias{autoSchema}
\title{class wraping xml string and its parameters}
\usage{
autoSchema(object, obj_xml_document)
}
\description{
Default schema based on output of C# DataTable serialisation with CreateXML method
}
\section{Slots}{

\describe{
\item{\code{xml_schema}}{"NullableNode"}

\item{\code{xml_body}}{"NullableNode"}

\item{\code{schema_tag}}{"character"}

\item{\code{schema_element}}{"character"}

\item{\code{body_tag}}{"character"}
}}

