% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\name{checkUnique}
\alias{checkUnique}
\title{Check if new incoming data has unique row}
\usage{
checkUnique(object, data)
}
\arguments{
\item{object}{object of class 'DataSet'.}

\item{data}{data.frame.}
}
\value{
\code{object} object of class 'DataSet'.
}
\description{
Check if new incoming data has unique row
}
\examples{
checkUnique(object,data)
}

