% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sockets.r
\docType{methods}
\name{closeConnection}
\alias{closeConnection}
\alias{closeConnection,ProcessSocket-method}
\title{Close socket connection}
\usage{
closeConnection(object)

\S4method{closeConnection}{ProcessSocket}(object)
}
\arguments{
\item{object}{object of class 'ProcessSocket'.}
}
\value{
\code{object} object of class 'ProcessSocket'.

\code{object} object of class 'ProcessSocket'.
}
\description{
Close socket connection
}
\section{Methods (by class)}{
\itemize{
\item \code{ProcessSocket}: Close socket connection
}}

