% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_analysis_objectstore.r
\docType{methods}
\name{commitAnalysisStore}
\alias{commitAnalysisStore}
\alias{commitAnalysisStore,AnalysisObjectStore-method}
\title{Commit AnalysisObjecstore to nonvolatile memory}
\usage{
commitAnalysisStore(object)

\S4method{commitAnalysisStore}{AnalysisObjectStore}(object)
}
\arguments{
\item{object}{object of class "AnalysisObjectStore"}
}
\value{
\code{object} object of class "AnalysisObjectStore"
}
\description{
Commit AnalysisObjecstore to nonvolatile memory
}
\section{Methods (by class)}{
\itemize{
\item \code{AnalysisObjectStore}: Commit AnalysisObjecstore to nonvolatile memory
}}

