% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_daily_riskmodel_objectstore.r
\docType{methods}
\name{commitDailyRiskModelObjectStore}
\alias{commitDailyRiskModelObjectStore}
\alias{commitDailyRiskModelObjectStore,DailyRiskModelObjectStore-method}
\title{Commit object to store and save}
\usage{
commitDailyRiskModelObjectStore(object)

\S4method{commitDailyRiskModelObjectStore}{DailyRiskModelObjectStore}(object)
}
\arguments{
\item{object}{object of class "DailyRiskModelObjectStore"}
}
\value{
\code{object} object of class "DailyRiskModelObjectStore"
}
\description{
Commit object to store and save
}
\section{Methods (by class)}{
\itemize{
\item \code{DailyRiskModelObjectStore}: Commit object to store and save
}}

