% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_ppmodel_objectstore.r
\name{commitPPModelStore}
\alias{commitPPModelStore}
\title{Commit store data to file}
\usage{
commitPPModelStore(object)
}
\arguments{
\item{object}{object of class "PPModelObjectStore"}
}
\value{
\code{object} object of class "PPModelObjectStore"
}
\description{
Saves data stored in the object into file.
}

