% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_trades_remote_objectstore.r
\docType{methods}
\name{commitTradeStore}
\alias{commitTradeStore}
\alias{commitTradeStore,TradeObjectStore-method}
\title{Commit store data to file}
\usage{
commitTradeStore(object)

\S4method{commitTradeStore}{TradeObjectStore}(object)
}
\arguments{
\item{object}{object of class "TradeObjectStore"}
}
\value{
\code{object} object of class "TradeObjectStore"

\code{object} object of class "TradeObjectStore"
}
\description{
Saves data stored in the object into file.
}
\section{Methods (by class)}{
\itemize{
\item \code{TradeObjectStore}: Commit store data to file
}}

