% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\name{compareUnique}
\alias{compareUnique}
\title{Check if new data is unique values in reference to stored data}
\usage{
compareUnique(object, new_data)
}
\arguments{
\item{object}{object of class 'DataSet'.}

\item{new_data}{data.frame with new data}
}
\value{
\code{new_data} data.frame with incoming_data without duplicated values
}
\description{
Check if new data is unique values in reference to stored data
}
\examples{
compareUnique(object,new_data)
}

