% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_composite_warehouse.r
\docType{class}
\name{CompositeWarehouse-class}
\alias{CompositeWarehouse-class}
\title{An S4 for composite Warehouse.}
\description{
Stores information about trades from multiple
warehouses
}
\section{Slots}{

\describe{
\item{\code{warehouse_names}}{"character"}

\item{\code{orig_trader_ids}}{"integer"}

\item{\code{orig_instruments}}{"list"}

\item{\code{orig_positions}}{"list"}

\item{\code{orig_psn_smries}}{"list"}

\item{\code{orig_dates}}{"list"}

\item{\code{orig_trade_ids}}{"list"}

\item{\code{orig_pads}}{"list"}

\item{\code{orig_maps}}{"list"}

\item{\code{orig_dtr_stores}}{"list"}
}}

