% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_ppmodel_library.r
\docType{data}
\name{context_meta}
\alias{context_meta}
\title{List of context meta features}
\format{An object of class \code{character} of length 1.}
\usage{
context_meta
}
\description{
List of context meta features
}
\keyword{datasets}

