% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_ppmodel_library.r
\docType{data}
\name{context_position_age_features}
\alias{context_position_age_features}
\title{List of position age features}
\format{An object of class \code{character} of length 3.}
\usage{
context_position_age_features
}
\description{
List of position age features
}
\keyword{datasets}

