% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_ppmodel_library.r
\docType{data}
\name{context_price_features}
\alias{context_price_features}
\title{List of context price features}
\format{An object of class \code{character} of length 21.}
\usage{
context_price_features
}
\description{
List of context price features
}
\keyword{datasets}

