% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_ppmodel_library.r
\docType{data}
\name{context_primary_placing_features}
\alias{context_primary_placing_features}
\title{List of primary placing features}
\format{An object of class \code{character} of length 2.}
\usage{
context_primary_placing_features
}
\description{
List of primary placing features
}
\keyword{datasets}

