% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_ppmodel_library.r
\docType{data}
\name{context_resultsday_features}
\alias{context_resultsday_features}
\title{List context results day features}
\format{An object of class \code{character} of length 3.}
\usage{
context_resultsday_features
}
\description{
List context results day features
}
\keyword{datasets}

