% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_ppmodel_library.r
\docType{data}
\name{context_volume_features}
\alias{context_volume_features}
\title{List of outcome volume features}
\format{An object of class \code{character} of length 1.}
\usage{
context_volume_features
}
\description{
List of outcome volume features
}
\keyword{datasets}

