% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_ppmodel_library.r
\docType{data}
\name{control_features}
\alias{control_features}
\title{List of control features}
\format{An object of class \code{character} of length 2.}
\usage{
control_features
}
\description{
List of control features
}
\keyword{datasets}

