% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_ppmodel_library.r
\docType{data}
\name{control_price_features}
\alias{control_price_features}
\title{List of control price features}
\format{An object of class \code{character} of length 3.}
\usage{
control_price_features
}
\description{
List of control price features
}
\keyword{datasets}

