% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urldatareader_functions.r
\name{convert2Matrix}
\alias{convert2Matrix}
\title{Convert xml document body to matrix}
\usage{
convert2Matrix(columns, types, obj_xml_document_body)
}
\arguments{
\item{columns}{character vector with column names of the new matrix}

\item{types}{character vector with classes of output columns}

\item{obj_xml_document_body}{character body of XML document}
}
\value{
\code{rval} matrix with new data from converted xml
}
\description{
Conversion of XMLDocument to a matrix given specified and column
names assumes that data is not ragged, is 2d and is uniformly numeric.
}

