% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_module.r
\docType{methods}
\name{createAnalysisModule}
\alias{createAnalysisModule}
\alias{createAnalysisModule,AnalysisModuleFactory-method}
\title{Method to create AnalysisModules}
\usage{
createAnalysisModule(object, key_func)

\S4method{createAnalysisModule}{AnalysisModuleFactory}(object, key_func)
}
\arguments{
\item{object}{object of class "AnalysisModuleFactory"}

\item{key_func}{"function" key generator function}
}
\value{
\code{rval} object of class derived from "VirtualAnalysisModule"

\code{rval} object of class derived from "VirtualAnalysisModule"
}
\description{
Method to create AnalysisModules
}
\section{Methods (by class)}{
\itemize{
\item \code{AnalysisModuleFactory}: Method to create AnalysisModules
}}

