% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\name{createIndex}
\alias{createIndex}
\title{Create new index for dataset}
\usage{
createIndex(object)
}
\arguments{
\item{object}{object of class 'DataSet'.}
}
\value{
\code{object} object of class 'DataSet'.
}
\description{
Create new index for dataset
}

