% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_daily_riskmodel_objectstore.r
\docType{class}
\name{DailyRiskModelObjectStore-class}
\alias{DailyRiskModelObjectStore-class}
\title{An S4 class for storing daily risk models.}
\description{
An S4 class for storing daily risk models.
}
\section{Slots}{

\describe{
\item{\code{components}}{"character"}

\item{\code{risk_model_q}}{"DailyRiskModelQuery"}
}}

