% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_daily_riskmodel_objectstore.r
\docType{class}
\name{DailyRiskModelQuery-class}
\alias{DailyRiskModelQuery-class}
\title{An S4 class handling queries to Daily Risk Model Objectstore.}
\description{
An S4 class handling queries to Daily Risk Model Objectstore.
}

