% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access_sql_procedures.r
\docType{class}
\name{DataAccess.SQLProcedureCall.Strategy_SelectAll-class}
\alias{DataAccess.SQLProcedureCall.Strategy_SelectAll-class}
\title{DataAccess.SQLProcedureCall.Strategy_SelectAll" class}
\description{
Implements handling querries for raw strategy data
}
\details{
Inherits from "VirtualSQLProcedureCall"
}

