% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_dataplex.r
\docType{data}
\name{dataplex_created}
\alias{dataplex_created}
\title{indicates if dataplex has been initialized}
\format{An object of class \code{logical} of length 1.}
\usage{
dataplex_created
}
\description{
indicates if dataplex has been initialized
}
\keyword{datasets}

