% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\docType{class}
\name{DataSet-class}
\alias{DataSet-class}
\title{An S4 class to represent generic data set.}
\description{
An S4 class to represent generic data set.
}
\section{Slots}{

\describe{
\item{\code{key_cols}}{"character"}

\item{\code{data_cols}}{"character"}

\item{\code{data}}{"data.frame"}

\item{\code{indexed}}{"logical"}

\item{\code{last_result}}{"data.frame"}

\item{\code{pad}}{"data.frame"}

\item{\code{unique_rows}}{"logical"}
}}

