% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\name{dataset_factory}
\alias{dataset_factory}
\title{Builds a dataset from a dataframe}
\usage{
dataset_factory(key_cols, data)
}
\arguments{
\item{data}{data.frame to generate dataset.from}

\item{key_col}{character vector ofs column names to be used as identifying keys.}
}
\value{
\code{object} object of class 'DataSet'
}
\description{
Builds a dataset from a dataframe
}
\examples{
rowCoalesceOn(object,by=NULL,fn=first)
}

