% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore.r
\docType{class}
\name{DataStore-class}
\alias{DataStore-class}
\title{An S4 class to represent generic data set.}
\description{
DataStore maintains a DataSet with an associated source URL
such thatdata already obtained is cached
}
\section{Slots}{

\describe{
\item{\code{dataset}}{"DataSet"}

\item{\code{urlparser}}{"URLParser"}

\item{\code{urlquery}}{"URLQuery"}

\item{\code{key_map}}{"KeyMap"}

\item{\code{last_result}}{"NullableResult"}
}}

