% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_sql.r
\docType{class}
\name{DataStore.SQL-class}
\alias{DataStore.SQL-class}
\title{An S4 class to represent generic data set.}
\description{
DataStore maintains a DataSet with an associated source SQL
procedure call
such thatdata already obtained is cached
}
\details{
Inherits from: "VirtualDataStore"
               "VirtualSQLQueryHandler"
}

