% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_sql.r
\docType{class}
\name{DataStore.VirtualSQLProcedureCall-class}
\alias{DataStore.VirtualSQLProcedureCall-class}
\title{Virtual S4 class handling sql procedures calls to DataStore DB.}
\description{
Implements handling of access to data via sql stored procedures calls.
Inherits from "VirtualSQLProcedureCall"
}

