% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keymap.r
\name{date_id_kgen_fn}
\alias{date_id_kgen_fn}
\title{Generate keys based on ID and Date columns}
\usage{
date_id_kgen_fn(x, cols, id_outname = NULL)
}
\arguments{
\item{x}{data.frame with columns based on which the keys will be generated}

\item{cols, }{cols[1] is the name of the ID column, cols[2] is the name of the date column}

\item{id_outname}{is the output name of the ID}
}
\value{
\code{key_values} data.frame with generated keys
}
\description{
Generates the values for each field to be retrieved from the data layer
also maps the URL field names to the XML variable names, output
data.frame column names should match the names of the query
fields in the middle layer URL
}

