% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_key_library.r
\name{date_range}
\alias{date_range}
\title{Generate key for date range}
\usage{
date_range(trader, start, end)
}
\arguments{
\item{trader}{integer, the user id}

\item{start}{Date, the most distant time (included, i.e. not upto this date, but inclusive)}

\item{end}{Date, the most recent time (included, i.e. not upto this date, but inclusive)}
}
\value{
\code{keys} data.frame with generated keys.
}
\description{
Generate key for range of dates
}

