% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_key_library.r
\name{dated_full_month}
\alias{dated_full_month}
\title{Generate key for full month of specific date}
\usage{
dated_full_month(trader, date)
}
\arguments{
\item{trader}{integer, the user id}

\item{date}{Date, the most recent time (included, i.e. not upto this date, but inclusive)}
}
\value{
\code{keys} data.frame with generated key.
}
\description{
Generate one month key for user.
}

