% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_key_library.r
\name{dated_this_year}
\alias{dated_this_year}
\title{Generate keys for this whole year lookback from begining to date}
\usage{
dated_this_year(trader, date)
}
\arguments{
\item{trader}{integer, the user id}

\item{date}{Date, the most recent time (included, i.e. not upto this date, but inclusive)}
}
\value{
\code{keys} data.frame with generated keys.
}
\description{
Generate key for whole year lookback from begining to the date month
}

