% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_key_library.r
\name{dated_three_year_lookback}
\alias{dated_three_year_lookback}
\title{Generate keys for whole three years lookback including date year}
\usage{
dated_three_year_lookback(trader, date)
}
\arguments{
\item{trader}{integer, the user id}

\item{date}{Date, the most recent time (included, i.e. not upto this date, but inclusive)}
}
\value{
\code{keys} data.frame with generated keys.
}
\description{
Generate keys for user with one year interval and three year lookback.
}

