% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_key_library.r
\name{dated_twelve_monthly_lookback}
\alias{dated_twelve_monthly_lookback}
\title{Generate keys for twelve months lookback for specific date}
\usage{
dated_twelve_monthly_lookback(trader, date)
}
\arguments{
\item{trader}{integer, the user id}

\item{date}{Date, the most recent time (included, i.e. not upto this date, but inclusive)}
}
\value{
\code{keys} data.frame with generated keys.
}
\description{
Generate twelve months lookback keys for user with
last day of previous month with regards to date as last date
and one month interval.
}

