% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_key_library.r
\name{dated_whole_year_lookback}
\alias{dated_whole_year_lookback}
\title{Generate key for whole year lookback from begining to end of date's year}
\usage{
dated_whole_year_lookback(trader, date)
}
\arguments{
\item{trader}{integer, the user id}

\item{date}{Date, the most recent time (included, i.e. not upto this date, but inclusive)}
}
\value{
\code{keys} data.frame with generated keys.
}
\description{
Generate key for whole year lookback from begining to end of date's year
}

