% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TE.DataAccess.r
\docType{package}
\name{foo}
\alias{foo}
\alias{foo-package}
\title{TE.DataAccess: Trading Enhancement Data access, cashing and preprocessing}
\description{
The TE.DataAccess package providing classes and functions for data retrieval,
caching  and manipulation for Trading Enhancement R engine
}
\section{TE.DataAccess classes}{

DataSet
}

