% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_analysis_objectstore.r
\name{get_analysis_objectstore_name}
\alias{get_analysis_objectstore_name}
\title{get analysis objectstore for keys and trader}
\usage{
get_analysis_objectstore_name(keys, trader_col = "TraderID")
}
\arguments{
\item{keys}{"data.frame" with query keys}

\item{trader_col}{"character" name of the column in key to be used as
identifier. Default is 'TraderID'}
}
\value{
\code{rv} "character" hashed name of the objectstore
}
\description{
get analysis objectstore for keys and trader
}

