% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_daily_riskmodel_objectstore.r
\name{get_earliest_model_objectstore}
\alias{get_earliest_model_objectstore}
\title{Get earliest risk model store}
\usage{
get_earliest_model_objectstore(model_prefix, lookback = 150)
}
\arguments{
\item{model_prefix}{character}

\item{lookback}{integer lookback of model in days}
}
\value{
\code{rm_str} if risk model objectstore found returns
object of class "DailyRiskModelObjectStore" otherwise NULL
}
\description{
Finds earliest risk model objectstore containg model
on querried date or earlier
}

