% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_daily_riskmodel_objectstore.r
\name{get_most_recent_model_objectstore}
\alias{get_most_recent_model_objectstore}
\title{Get most recent risk model store}
\usage{
get_most_recent_model_objectstore(model_prefix, date = today() - 1,
  lookback = 150L)
}
\arguments{
\item{model_prefix}{character}

\item{date}{Date date for model}

\item{lookback}{integer lookback of model in days}
}
\value{
\code{rm_str} if risk model objectstore found returns
object of class "DailyRiskModelObjectStore" otherwise NULL
}
\description{
Finds most recent risk model objectstore containg model
on querried date or earlier
}

