% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_ppmodel_objectstore.r
\name{get_ppmodel_objectstore_name}
\alias{get_ppmodel_objectstore_name}
\title{Generate PPModel objectstore name from keys}
\usage{
get_ppmodel_objectstore_name(keys)
}
\arguments{
\item{keys}{"data.frame", keys from which name(s) of objectstore is created}
}
\description{
Creates PPModel ObjectStore and loads data from
associated file if exists.
}

