% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_trades_remote_objectstore.r
\name{get_trade_objectstore_name}
\alias{get_trade_objectstore_name}
\title{Generate Trade objectstore name from keys}
\usage{
get_trade_objectstore_name(keys)
}
\arguments{
\item{keys}{"data.frame", keys from which name(s) of objectstore is created}
}
\description{
Creates Trade ObjectStore and loads data from
associated file if exists.
}

