% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_trades_remote_objectstore.r
\docType{methods}
\name{getAllTradesdFromTradeStore}
\alias{getAllTradesdFromTradeStore}
\alias{getAllTradesdFromTradeStore,TradeObjectStore-method}
\title{Query store for Trade}
\usage{
getAllTradesdFromTradeStore(object)

\S4method{getAllTradesdFromTradeStore}{TradeObjectStore}(object)
}
\arguments{
\item{object}{object of class "TradeObjectStore"}
}
\value{
\code{rval} object of class "Trade"

\code{rval} list of objects of class "Trade"
}
\description{
Querries Trade Objectstore for Trade stored under given key
Returns Trade if present NULL otherwise
}
\section{Methods (by class)}{
\itemize{
\item \code{TradeObjectStore}: Query store for Trade

Querries Trade Objectstore for Trade stored under given key
Returns Trade if present NULL otherwise
}}

