% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sockets.r
\docType{methods}
\name{getData,ProcessSocket-method}
\alias{getData,ProcessSocket-method}
\title{Get data read from socket}
\usage{
\S4method{getData}{ProcessSocket}(object)
}
\arguments{
\item{object}{object of class 'ProcessSocket'.}
}
\value{
\code{data} character, data read from socket..
}
\description{
Retrieve data previously read from socket using
\code{readConnection(object)}
}
\examples{
getData(object)
}

