% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\name{getData}
\alias{getData}
\title{Return data stored in the DataSet}
\usage{
getData(object)
}
\arguments{
\item{object}{object of class 'DataSet'.}
}
\value{
\code{data} data.frame with raw data.
}
\description{
Return data stored in the DataSet
}

