% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sockets.r
\docType{methods}
\name{getDataFromSocket}
\alias{getDataFromSocket}
\alias{getDataFromSocket,ProcessSocket-method}
\title{Get data read from socket}
\usage{
getDataFromSocket(object)

\S4method{getDataFromSocket}{ProcessSocket}(object)
}
\arguments{
\item{object}{object of class 'ProcessSocket'.}
}
\value{
\code{data} character, data read from socket..
}
\description{
Retrieve data previously read from socket using
\code{readConnection(object)}
}
\section{Methods (by class)}{
\itemize{
\item \code{ProcessSocket}: Get data read from socket

Retrieve data previously read from socket using
\code{readConnection(object)}
}}

