% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\docType{methods}
\name{getDataSetDataColumnNames}
\alias{getDataSetDataColumnNames}
\alias{getDataSetDataColumnNames,DataSet-method}
\title{Return column names of data columns}
\usage{
getDataSetDataColumnNames(object)

\S4method{getDataSetDataColumnNames}{DataSet}(object)
}
\arguments{
\item{object}{object of class 'DataSet'.}
}
\value{
\code{key_cols} "character" vector with key column names
}
\description{
Return column names of data columns
}
\section{Methods (by class)}{
\itemize{
\item \code{DataSet}: Sets new data in the DataSet

Return column names of data columns
}}

