% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\docType{methods}
\name{getDataSetKeyColumnNames}
\alias{getDataSetKeyColumnNames}
\alias{getDataSetKeyColumnNames,DataSet-method}
\title{Return column names of key columns}
\usage{
getDataSetKeyColumnNames(object)

\S4method{getDataSetKeyColumnNames}{DataSet}(object)
}
\arguments{
\item{object}{object of class 'DataSet'.}
}
\value{
\code{data_cols} "character" vector with key column names
}
\description{
Return column names of key columns
}
\section{Methods (by class)}{
\itemize{
\item \code{DataSet}: Sets new data in the DataSet

Return column names of key columns
}}

