% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_daily_riskmodel_objectstore.r
\docType{methods}
\name{getEarliestRiskModelDate}
\alias{getEarliestRiskModelDate}
\alias{getEarliestRiskModelDate,DailyRiskModelObjectStore,character,integer-method}
\title{Get earliest risk model date stored in object}
\usage{
getEarliestRiskModelDate(object, name, lookback = 150L)


  \S4method{getEarliestRiskModelDate}{DailyRiskModelObjectStore,character,integer}(object,
  name, lookback = 150)
}
\arguments{
\item{object}{object of class "DailyRiskModelObjectStore"}

\item{name}{"character" name of the model eg: "developed_europe_prototype"}

\item{lookback}{"integer" number of lookback horizon of the model
Method will return latest model date not later than this parameter if not NULL.
This is provided to be able to query gaps in model dates.}
}
\value{
\code{date} "Date" date of the most recent risk model.
Returns "NULL" if nothing is stored
or only components that are not model specific are stored.
}
\description{
Get earliest risk model date stored in object
}
\section{Methods (by class)}{
\itemize{
\item \code{object = DailyRiskModelObjectStore,name = character,lookback = integer}: Get earliest risk model date stored in object
}}

