% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectstore.r
\docType{methods}
\name{getID}
\alias{getID}
\alias{getID,VirtualObjectStore-method}
\title{Get ID of the objectstore}
\usage{
getID(object)

\S4method{getID}{VirtualObjectStore}(object)
}
\arguments{
\item{object}{object of class "VirtualObjectStore"}
}
\value{
\code{id} "character" object id

\code{id} "character" object id
}
\description{
Get ID of the objectstore
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualObjectStore}: Get ID of the objectstore
}}

