% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_warehouse_objectstore.r
\name{getKeyMapID}
\alias{getKeyMapID}
\title{Get KeyMap ID}
\usage{
getKeyMapID(object)
}
\arguments{
\item{object}{object of class "WarehouseObjectStore"}
}
\value{
\code{keymap_id} "character"
}
\description{
Get KeyMap ID
}

