% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_daily_riskmodel_objectstore.r
\docType{methods}
\name{getMostRecentRiskModelDate}
\alias{getMostRecentRiskModelDate}
\alias{getMostRecentRiskModelDate,DailyRiskModelObjectStore,character,integer-method}
\title{Get most recent risk model date stored in object}
\usage{
getMostRecentRiskModelDate(object, name, lookback = 150L)


  \S4method{getMostRecentRiskModelDate}{DailyRiskModelObjectStore,character,integer}(object,
  name, lookback = 150)
}
\arguments{
\item{object}{object of class "DailyRiskModelObjectStore"}

\item{name}{"character" name of the model eg: "developed_europe_prototype"}

\item{lookback}{"integer" number of lookback horizon of the model}
}
\value{
\code{date} "Date" date of the most recent risk model.
Returns "NULL" if nothing is stored
or only components that are not model specific are stored.
}
\description{
Get most recent risk model date stored in object
}
\section{Methods (by class)}{
\itemize{
\item \code{object = DailyRiskModelObjectStore,name = character,lookback = integer}: Get most recent risk model date stored in object
}}

