% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectstore.r
\docType{methods}
\name{getObjectstoreKey}
\alias{getObjectstoreKey}
\alias{getObjectstoreKey,VirtualObjectStore-method}
\title{Get key of the objectstore}
\usage{
getObjectstoreKey(object)

\S4method{getObjectstoreKey}{VirtualObjectStore}(object)
}
\arguments{
\item{object}{object of class "VirtualObjectStore"}
}
\value{
\code{key} "data.frame" with query defining store
}
\description{
Get key of the objectstore
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualObjectStore}: Get ID of the objectstore
}}

