% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_warehouse_objectstore.r
\docType{methods}
\name{getObjectStoreKeyMap}
\alias{getObjectStoreKeyMap}
\alias{getObjectStoreKeyMap,WarehouseObjectStore-method}
\title{Get objectstore query object}
\usage{
getObjectStoreKeyMap(object)

\S4method{getObjectStoreKeyMap}{WarehouseObjectStore}(object)
}
\arguments{
\item{object}{object of class "WarehouseObjectStore"}
}
\value{
\code{key_map} "character" object of class "KeyMap"
}
\description{
Get objectstore query object
}
\section{Methods (by class)}{
\itemize{
\item \code{WarehouseObjectStore}: Get key map of the objectstore
}}

