% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectstore.r
\docType{methods}
\name{getObjectStoreQuery}
\alias{getObjectStoreQuery}
\alias{getObjectStoreQuery,VirtualObjectStore-method}
\title{Get objectstore query object}
\usage{
getObjectStoreQuery(object)

\S4method{getObjectStoreQuery}{VirtualObjectStore}(object)
}
\arguments{
\item{object}{object of class "VirtualObjectStore"}
}
\value{
\code{objectstore_q} "character" object of class "ObjectQuery"
}
\description{
Get objectstore query object
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualObjectStore}: Get ID of the objectstore
}}

