% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_warehouse_objectstore.r
\name{getQueryID}
\alias{getQueryID}
\title{Get QueryID ID}
\usage{
getQueryID(object)
}
\arguments{
\item{object}{object of class "WarehouseObjectStore"}
}
\value{
\code{objectquery_id} "character"
}
\description{
Get QueryID ID
}

