% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_trade_factory.r
\name{getRawPositionData}
\alias{getRawPositionData}
\title{Get position data from Warehouse}
\usage{
getRawPositionData(object)
}
\arguments{
\item{object}{object of class "TradeWarehouse"}
}
\value{
\code{positions} object of class "DataSet" with position data
}
\description{
Returns all position data stored in Warehouse
}

