% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_trade_factory.r
\docType{methods}
\name{getRawPositionSummary}
\alias{getRawPositionSummary}
\alias{getRawPositionSummary,TradeWarehouse-method}
\title{Get position summary data from Warehouse}
\usage{
getRawPositionSummary(object)

\S4method{getRawPositionSummary}{TradeWarehouse}(object)
}
\arguments{
\item{object}{object of class "TradeWarehouse"}
}
\value{
\code{psn_summary} object of class "DataSet" with position summary data

\code{psn_summary} object of class "DataSet" with position summary data
}
\description{
Returns all position summary data stored in Warehouse
}
\section{Methods (by class)}{
\itemize{
\item \code{TradeWarehouse}: Get position summary data from Warehouse

Returns all position summary data stored in Warehouse
}}

