% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_daily_riskmodel_objectstore.r
\docType{methods}
\name{getRiskModelComponentOnDate}
\alias{getRiskModelComponentOnDate}
\alias{getRiskModelComponentOnDate,DailyRiskModelObjectStore,character,character,Date,integer-method}
\title{Get most recent risk model date stored in object}
\usage{
getRiskModelComponentOnDate(object, name, component, date, lookback = 150)


  \S4method{getRiskModelComponentOnDate}{DailyRiskModelObjectStore,character,character,Date,integer}(object,
  name, component, date, lookback = 150)
}
\arguments{
\item{object}{object of class "DailyRiskModelObjectStore"}

\item{name}{"character" name of the model eg: "developed_europe_prototype"}

\item{component}{"character" name of the component of the risk model, possible values are:
c('ImpliedFactorReturns', 'ResidualReturns', 'Betas',
  'FactorCorrelation', 'FactorVariance', 'MarketStyle')}

\item{date}{"Date" date of the risk model for which it was computed.}

\item{lookback}{"integer" number of lookback horizon of the model}
}
\value{
\code{rdata} "data.frame with risk component
}
\description{
Get most recent risk model date stored in object
}
\section{Methods (by class)}{
\itemize{
\item \code{object = DailyRiskModelObjectStore,name = character,component = character,date = Date,lookback = integer}: Get most recent risk model date stored in object
}}

