% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_daily_riskmodel_objectstore.r
\docType{methods}
\name{getRiskModelComponents}
\alias{getRiskModelComponents}
\alias{getRiskModelComponents,DailyRiskModelObjectStore-method}
\title{Get list of storable components}
\usage{
getRiskModelComponents(object)

\S4method{getRiskModelComponents}{DailyRiskModelObjectStore}(object)
}
\arguments{
\item{object}{object of class "DailyRiskModelObjectStore"}
}
\value{
\code{components} "character" vector of storable components
}
\description{
Get list of storable components
}
\section{Methods (by class)}{
\itemize{
\item \code{DailyRiskModelObjectStore}: Get list of storable components
}}

