% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_warehouse_objectstore.r
\docType{methods}
\name{getWarehouseFromStore}
\alias{getWarehouseFromStore}
\alias{getWarehouseFromStore,WarehouseObjectStore-method}
\title{Get warehouse from WarehouseObjectstore for given keys}
\usage{
getWarehouseFromStore(object, trader_id, start, end)

\S4method{getWarehouseFromStore}{WarehouseObjectStore}(object, trader_id, start,
  end)
}
\arguments{
\item{object}{object of class "WarehouseObjectStore"}

\item{trader_id}{"integer", trader ID}

\item{start}{"Date" start date of Warehouse}

\item{end}{"Date" end date of Warehouse}
}
\value{
\code{wh} object of class "CompositeWarehouse" if present otherwise NULL

\code{wh} object of class "CompositeWarehouse" if present otherwise NULL
}
\description{
Retrieves from WarehouseObjectstore warehouse matching
keys passed as arguments.
}
\section{Methods (by class)}{
\itemize{
\item \code{WarehouseObjectStore}: Get warehouse from WarehouseObjectstore for given keys

Retrieves from WarehouseObjectstore warehouse matching
keys passed as arguments.
}}

