% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\name{hashRows}
\alias{hashRows}
\title{Hash keys of new incomming rows}
\usage{
hashRows(object, rows, keys, type = "local")
}
\arguments{
\item{object}{object of class 'DataSet'.}

\item{rows}{data.frame with new rows to be added to dataset}

\item{keys}{columnames of columns used as keys for hashing}

\item{type}{= 'local', c('foreign', 'local'),}
}
\value{
\code{hashed_rows} vector of hashed key values
}
\description{
Hash keys of new incomming rows
}

