% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\name{initialiseOrAppendData}
\alias{initialiseOrAppendData}
\title{Appends new data in the dataset or initializes it if not done}
\usage{
initialiseOrAppendData(object, data)
}
\arguments{
\item{object}{object of class 'DataSet'.}

\item{data}{data.frame.}
}
\value{
\code{object} object of class 'DataSet'.
}
\description{
Appends new data in the dataset or initializes it if not done
}
\examples{
initialiseOrAppendData(object,data)
}

