% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\name{initialiseOrSetData}
\alias{initialiseOrSetData}
\title{Sets new data in the dataset and initializes it if not done}
\usage{
initialiseOrSetData(object, data)
}
\arguments{
\item{object}{object of class 'DataSet'.}

\item{data}{data.frame.}
}
\value{
\code{object} object of class 'DataSet'.
}
\description{
Sets new data in the dataset and initializes it if not done
}

