% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_ppmodel_objectstore.r
\name{initialisePPModelStore}
\alias{initialisePPModelStore}
\title{#' Initialize method for "PPModelObjectStore" class
#'
#' @param .Object, object of class "PPModelObjectStore"
#' @param id id to set when initializing
#' @return \code{.Object} object of class "PPModelObjectStore"}
\usage{
initialisePPModelStore(object)
}
\description{
setMethod("initialize", "PPModelObjectStore",
          function(.Object,id){
            .Object@id <- id
            .Object
          }
)
}

