% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_composite_warehouse.r
\docType{methods}
\name{initialize,CompositeWarehouse-method}
\alias{initialize,CompositeWarehouse-method}
\title{Initialize method for "CompositeWarehouse" class}
\usage{
\S4method{initialize}{CompositeWarehouse}(.Object)
}
\arguments{
\item{.Object, }{object of class "CompositeWarehouse"}
}
\value{
\code{.Object} object of class "CompositeWarehouse"
}
\description{
Prototype is made only once and then copied with each call to 'new'
for objects along the same branch in the class hierachy.
Must ensure new environments created since they are passed
by reference.
}

