% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_dataplex.r
\docType{methods}
\name{initialize,DataPlex-method}
\alias{initialize,DataPlex-method}
\title{Initialization method for DataPlex
Populates DataPlex with respective DataStores}
\usage{
\S4method{initialize}{DataPlex}(.Object)
}
\arguments{
\item{.Object}{object of class "DataPlex"}
}
\value{
\code{.Object} object of class "DataPlex"
}
\description{
Initialization method for DataPlex
Populates DataPlex with respective DataStores
}

