% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_RAIDdata.r
\docType{methods}
\name{initialize,DealingBookURL-method}
\alias{initialize,DealingBookURL-method}
\title{Initialize method for "DealingBookURL" class}
\usage{
\S4method{initialize}{DealingBookURL}(.Object, trader_id, start, end)
}
\arguments{
\item{.Object, }{object of class "DealingBookURL"}

\item{trader_id}{"integer" trader id}

\item{start}{"Date" start date}

\item{end}{"Date" end date}
}
\value{
\code{.Object} object of class "DealingBookURL"
}
\description{
Initialize method for "DealingBookURL" class
}

