% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_RAIDdata.r
\docType{methods}
\name{initialize,DynamicHistoryURL-method}
\alias{initialize,DynamicHistoryURL-method}
\title{Initialize method for "DynamicHistoryURL" class}
\usage{
\S4method{initialize}{DynamicHistoryURL}(.Object, instrument, start, end)
}
\arguments{
\item{.Object, }{object of class "DynamicHistoryURL"}

\item{instrument}{"integer" instrument ID}

\item{start}{"Date" start date}

\item{end}{"Date" end date}
}
\value{
\code{.Object} object of class "DynamicHistoryURL"
}
\description{
Initialize method for "DynamicHistoryURL" class
}

