% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_RAIDdata.r
\docType{methods}
\name{initialize,EventsTypesURL-method}
\alias{initialize,EventsTypesURL-method}
\title{Initialize method for "EventsTypesURL" class}
\usage{
\S4method{initialize}{EventsTypesURL}(.Object)
}
\arguments{
\item{.Object, }{object of class "EventsTypesURL"}
}
\value{
\code{.Object} object of class "EventsTypesURL"
}
\description{
Initialize method for "EventsTypesURL" class
}

