% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_RAIDdata.r
\docType{methods}
\name{initialize,EventsURL-method}
\alias{initialize,EventsURL-method}
\title{Initialize method for "EventsURL" class}
\usage{
\S4method{initialize}{EventsURL}(.Object, instrument, start, end)
}
\arguments{
\item{.Object, }{object of class "EventsURL"}

\item{instrument}{"integer" instrument ID}

\item{start}{"Date" start date}

\item{end}{"Date" end date}
}
\value{
\code{.Object} object of class "EventsURL"
}
\description{
Initialize method for "EventsURL" class
}

