% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_RAIDdata.r
\docType{methods}
\name{initialize,ExtendedPositionDataURL-method}
\alias{initialize,ExtendedPositionDataURL-method}
\title{Initialize method for "ExtendedPositionDataURL" class}
\usage{
\S4method{initialize}{ExtendedPositionDataURL}(.Object, strategy, start, end)
}
\arguments{
\item{.Object, }{object of class "ExtendedPositionDataURL"}

\item{strategy}{"character" strategy name}

\item{start}{"Date" start date}

\item{end}{"Date" end date}
}
\value{
\code{.Object} object of class "ExtendedPositionDataURL"
}
\description{
Initialize method for "ExtendedPositionDataURL" class
}

