% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_RAIDdata.r
\docType{methods}
\name{initialize,InstrumentHistoryURL-method}
\alias{initialize,InstrumentHistoryURL-method}
\title{Initialize method for "InstrumentHistoryURL" class}
\usage{
\S4method{initialize}{InstrumentHistoryURL}(.Object, instrument_ids, start, end)
}
\arguments{
\item{.Object, }{object of class "InstrumentHistoryURL"}

\item{instrument_ids}{"integer" vector of instrument ID's}

\item{start}{"Date" start date}

\item{end}{"Date" end date}
}
\value{
\code{.Object} object of class "InstrumentHistoryURL"
}
\description{
Initialize method for "InstrumentHistoryURL" class
}

