% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_configs.r
\docType{methods}
\name{initialize,ModelDefaults-method}
\alias{initialize,ModelDefaults-method}
\title{Initialize method for "ModelDefaults" class}
\usage{
\S4method{initialize}{ModelDefaults}(.Object)
}
\arguments{
\item{.Object, }{object of class "ModelDefaults"}
}
\value{
\code{.Object} object of class "ModelDefaults"
}
\description{
Initialize method for "ModelDefaults" class
}

