% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_RAIDdata.r
\docType{methods}
\name{initialize,PositionDiaryURL-method}
\alias{initialize,PositionDiaryURL-method}
\title{Initialize method for "PositionDiaryURL" class}
\usage{
\S4method{initialize}{PositionDiaryURL}(.Object, trader_id, start, end)
}
\arguments{
\item{.Object, }{object of class "PositionDiaryURL"}

\item{trader_id}{"integer" trader id}

\item{start}{"Date" start date}

\item{end}{"Date" end date}
}
\value{
\code{.Object} object of class "PositionDiaryURL"
}
\description{
Initialize method for "PositionDiaryURL" class
}

