% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_RAIDdata.r
\docType{methods}
\name{initialize,PositionHistoryURL-method}
\alias{initialize,PositionHistoryURL-method}
\title{Initialize method for "PositionHistoryURL" class}
\usage{
\S4method{initialize}{PositionHistoryURL}(.Object, user_id, start, end)
}
\arguments{
\item{.Object, }{object of class "PositionHistoryURL"}

\item{user_id}{"integer" trader ID}

\item{start}{"Date" start date}

\item{end}{"Date" end date}
}
\value{
\code{.Object} object of class "PositionHistoryURL"
}
\description{
Initialize method for "PositionHistoryURL" class
}

