% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_ppmodel.r
\docType{methods}
\name{initialize,PPModel-method}
\alias{initialize,PPModel-method}
\title{Initialize method for "PPModel" class}
\usage{
\S4method{initialize}{PPModel}(.Object, keys)
}
\arguments{
\item{.Object, }{object of class "PPModel"}

\item{keys}{"character" key names for ppmodel dataset and query}
}
\value{
\code{.Object} object of class "PPModel"
}
\description{
Initialize method for "PPModel" class
}

