% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_ppmodel_objectstore.r
\docType{methods}
\name{initialize,PPModelObjectStore-method}
\alias{initialize,PPModelObjectStore-method}
\title{Initialize method for "PPModelObjectStore" class}
\usage{
\S4method{initialize}{PPModelObjectStore}(.Object, id)
}
\arguments{
\item{.Object, }{object of class "PPModelObjectStore"}

\item{id}{id to set when initializing}
}
\value{
\code{.Object} object of class "PPModelObjectStore"
}
\description{
Initialize method for "PPModelObjectStore" class
}

