% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_analysis_objectstore.r
\docType{methods}
\name{initialize,RemoteAnalysisQuery-method}
\alias{initialize,RemoteAnalysisQuery-method}
\title{Initialize method for "RemoteAnalysisQuery" class}
\usage{
\S4method{initialize}{RemoteAnalysisQuery}(.Object)
}
\arguments{
\item{.Object, }{object of class "RemoteAnalysisQuery"}
}
\value{
\code{.Object} object of class "RemoteAnalysisQuery"
}
\description{
Initialize method for "RemoteAnalysisQuery" class
}

