% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_objectstore.r
\docType{methods}
\name{initialize,RemoteObjectQuery-method}
\alias{initialize,RemoteObjectQuery-method}
\title{Initialize method for "RemoteObjectQuery" class}
\usage{
\S4method{initialize}{RemoteObjectQuery}(.Object)
}
\arguments{
\item{.Object, }{object of class "RemoteObjectQuery"}
}
\value{
\code{.Object} object of class "RemoteObjectQuery"
}
\description{
Initialize method for "RemoteObjectQuery" class
}

