% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_ppmodel_objectstore.r
\docType{methods}
\name{initialize,RemotePPModelQuery-method}
\alias{initialize,RemotePPModelQuery-method}
\title{Initialize method for "RemotePPModelQuery" class}
\usage{
\S4method{initialize}{RemotePPModelQuery}(.Object)
}
\arguments{
\item{.Object, }{object of class "RemotePPModelQuery"}
}
\value{
\code{.Object} object of class "RemotePPModelQuery"
}
\description{
Initialize method for "RemotePPModelQuery" class
}

