% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_trades_remote_objectstore.r
\docType{methods}
\name{initialize,RemoteTradeQuery-method}
\alias{initialize,RemoteTradeQuery-method}
\title{Initialize method for "RemoteTradeQuery" class}
\usage{
\S4method{initialize}{RemoteTradeQuery}(.Object)
}
\arguments{
\item{.Object, }{object of class "RemoteTradeQuery"}
}
\value{
\code{.Object} object of class "RemoteTradeQuery"
}
\description{
Initialize method for "RemoteTradeQuery" class
}

