% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_configs.r
\docType{methods}
\name{initialize,RiskModelDefaults-method}
\alias{initialize,RiskModelDefaults-method}
\title{Initialize method for "RiskModelDefaults" class}
\usage{
\S4method{initialize}{RiskModelDefaults}(.Object)
}
\arguments{
\item{.Object, }{object of class "RiskModelDefaults"}
}
\value{
\code{.Object} object of class "RiskModelDefaults"
}
\description{
Initialize method for "RiskModelDefaults" class
}

