% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_RAIDdata.r
\docType{methods}
\name{initialize,Static2HistoryURL-method}
\alias{initialize,Static2HistoryURL-method}
\title{Initialize method for "Static2HistoryURL" class}
\usage{
\S4method{initialize}{Static2HistoryURL}(.Object, instrument, start, end)
}
\arguments{
\item{.Object, }{object of class "Static2HistoryURL"}

\item{instrument}{"integer" instrument ID}

\item{start}{"Date" start date}

\item{end}{"Date" end date}
}
\value{
\code{.Object} object of class "Static2HistoryURL"
}
\description{
Initialize method for "Static2HistoryURL" class
}

