% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_RAIDdata.r
\docType{methods}
\name{initialize,StaticHistoryURL-method}
\alias{initialize,StaticHistoryURL-method}
\title{Initialize method for "StaticHistoryURL" class}
\usage{
\S4method{initialize}{StaticHistoryURL}(.Object, instrument, start, end)
}
\arguments{
\item{.Object, }{object of class "StaticHistoryURL"}

\item{instrument}{"integer" instrument ID}

\item{start}{"Date" start date}

\item{end}{"Date" end date}
}
\value{
\code{.Object} object of class "StaticHistoryURL"
}
\description{
Initialize method for "StaticHistoryURL" class
}

