% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_RAIDdata.r
\docType{methods}
\name{initialize,TradeHistoryURL-method}
\alias{initialize,TradeHistoryURL-method}
\title{Initialize method for "TradeHistoryURL" class}
\usage{
\S4method{initialize}{TradeHistoryURL}(.Object, user_ids, start, end)
}
\arguments{
\item{.Object, }{object of class "TradeHistoryURL"}

\item{user_ids}{"integer" trader ID}

\item{start}{"Date" start date}

\item{end}{"Date" end date}
}
\value{
\code{.Object} object of class "TradeHistoryURL"
}
\description{
Initialize method for "TradeHistoryURL" class
}

