% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_trades_remote_objectstore.r
\docType{methods}
\name{initialize,TradeObjectStore-method}
\alias{initialize,TradeObjectStore-method}
\title{Initialize method for "TradeObjectStore" class}
\usage{
\S4method{initialize}{TradeObjectStore}(.Object, id, key)
}
\arguments{
\item{.Object, }{object of class "TradeObjectStore"}

\item{id}{"character "id to set when initializing}

\item{key}{"data.frame" key defining objectstore to set when initializing}
}
\value{
\code{.Object} object of class "TradeObjectStore"
}
\description{
Initialize method for "TradeObjectStore" class
}

