% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_trade_factory.r
\docType{methods}
\name{initialize,TradeWarehouse-method}
\alias{initialize,TradeWarehouse-method}
\title{Initialize method for "TradeWarehouse" class}
\usage{
\S4method{initialize}{TradeWarehouse}(.Object)
}
\arguments{
\item{.Object, }{object of class "TradeWarehouse"}
}
\value{
\code{.Object} object of class "TradeWarehouse"
}
\description{
Initialize method for "TradeWarehouse" class
}

