% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectstore.r
\docType{methods}
\name{initialize,VirtualObjectStore-method}
\alias{initialize,VirtualObjectStore-method}
\title{Initialize method for "VirtualObjectStore" class}
\usage{
\S4method{initialize}{VirtualObjectStore}(.Object, id)
}
\arguments{
\item{.Object, }{object of class "VirtualObjectStore"}

\item{id}{"character" name of the objectstore}
}
\value{
\code{.Object} object of class "VirtualObjectStore"
}
\description{
Initialize method for "VirtualObjectStore" class
}

