% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_objectstore.r
\docType{methods}
\name{initialize,VirtualRemoteObjectStore-method}
\alias{initialize,VirtualRemoteObjectStore-method}
\title{Initialize method for "VirtualRemoteObjectStore" class}
\usage{
\S4method{initialize}{VirtualRemoteObjectStore}(.Object, id)
}
\arguments{
\item{.Object, }{object of class "VirtualRemoteObjectStore"}

\item{id}{"character" name of the objectstore}
}
\value{
\code{.Object} object of class "VirtualRemoteObjectStore"
}
\description{
Initialize method for "VirtualRemoteObjectStore" class
}

