% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_trade.r
\docType{methods}
\name{initialize,VirtualTrade-method}
\alias{initialize,VirtualTrade-method}
\title{Initialize method for "VirtualTrade" class}
\usage{
\S4method{initialize}{VirtualTrade}(.Object, leg_start, leg_end, trader,
  instrument, strategy, long, value_usd, consolidation, status)
}
\arguments{
\item{.Object, }{object of class "VirtualTrade"}
}
\value{
\code{.Object} object of class "VirtualTrade"
}
\description{
Initialize method for "VirtualTrade" class
}

