% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_configs.r
\docType{methods}
\name{initialize,VisualisationDefaults-method}
\alias{initialize,VisualisationDefaults-method}
\title{Initialize method for "VisualisationDefaults" class}
\usage{
\S4method{initialize}{VisualisationDefaults}(.Object)
}
\arguments{
\item{.Object, }{object of class "VisualisationDefaults"}
}
\value{
\code{.Object} object of class "VisualisationDefaults"
}
\description{
Initialize method for "VisualisationDefaults" class
}

