% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\name{innerJoin}
\alias{innerJoin}
\title{join existing data in dataset with incomming dataset
inner join might result in empty data}
\usage{
innerJoin(object, incoming_dataset, on_columns, aliases = NULL,
  joinmode = "inner")
}
\arguments{
\item{object}{object of class 'DataSet'.}

\item{incoming_dataset}{object of class 'DataSet'.}

\item{on_columns}{column names used for mering}

\item{aliases}{incomming data.frame column aliases of existing dataset names}

\item{joinmode}{join mode, default is 'inner'}
}
\value{
\code{object} object of class 'DataSet'.
}
\description{
join existing data in dataset with incomming dataset
inner join might result in empty data
}
\examples{
innerJoin(object,incoming_dataset,on_columns,aliases=NULL,joinmode='inner')
}

