% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\name{innerJoinFrame}
\alias{innerJoinFrame}
\title{Join existing data in dataset with incomming data
inner join might result in empty data}
\usage{
innerJoinFrame(object, frame, key_cols, aliases = NULL, joinmode = "inner")
}
\arguments{
\item{object}{object of class 'DataSet'.}

\item{frame}{object of class 'data.frame'.}

\item{key_cols}{column names used for merging}

\item{aliases}{incomming data.frame column aliases of existing dataset names}

\item{joinmode}{join mode, default is 'inner'}
}
\value{
\code{object} object of class 'DataSet'.
}
\description{
Join existing data in dataset with incomming data
inner join might result in empty data
}
\examples{
innerJoinFrame(object,frame,key_cols,aliases=NULL,joinmode='inner')
}

