% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_trades_remote_objectstore.r
\name{key_from_trade_objectstore_name}
\alias{key_from_trade_objectstore_name}
\title{helper function to generate key from objectstore name}
\usage{
key_from_trade_objectstore_name(name)
}
\arguments{
\item{name}{"character" name of the objectstore}
}
\value{
\code{key} "data.frame" with columns "id", "leg_start", "leg_end"
}
\description{
helper function to generate key from objectstore name
}

