% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_key_library.r
\name{key_generator}
\alias{dated_four_monthly_lookback}
\alias{dated_full_month}
\alias{dated_one_year_lookback}
\alias{dated_three_day_lookback}
\alias{dated_three_monthly_lookback}
\alias{dated_three_year_lookback}
\alias{dated_twelve_monthly_lookback}
\alias{dated_whole_year_lookback}
\alias{key_generator}
\alias{last_month}
\alias{last_year}
\alias{range_years_lookback}
\alias{this_year}
\alias{three_monthly_lookback}
\alias{three_year_lookback}
\title{Generate data frame of start and end dates for a user.}
\usage{
key_generator(trader, end, interval, lookback, calendar = TRUE,
  lookback_unit = "months", cols = c("id", "start", "end"))

three_monthly_lookback(trader)

last_month(trader)

dated_three_monthly_lookback(trader, date)

dated_four_monthly_lookback(trader, date)

dated_twelve_monthly_lookback(trader, date)

dated_three_day_lookback(trader, date)

dated_full_month(trader, date)

dated_one_year_lookback(trader, date)

dated_three_year_lookback(trader, date)

dated_whole_year_lookback(trader, date)

three_year_lookback(trader)

last_year(trader)

this_year(trader)

range_years_lookback(trader, start, end)
}
\arguments{
\item{trader}{integer, the user id}

\item{end}{Date, the most recent time (included, i.e. not upto this date, but inclusive)}

\item{interval}{integer, the date decrement}

\item{lookback}{integer, the number of decrements including the end date}

\item{calendar}{logical, if calendar is true, it is assumed that end is the final day of a
calendar period (eg. month,week) and the date intervals are adjusted to span the calendar interval.}

\item{lookback_unit}{character, name of time unit as recognized by lubridate:
c("days", "weeks", "months", "years"), default is "months"}

\item{cols}{character, name output columns, default is c('id','start','end')}

\item{date}{Date, the most recent time (included, i.e. not upto this date, but inclusive)}

\item{start}{Date, the most distant time (included, i.e. not upto this date, but inclusive)}

\item{trader}{integer, the user id}

\item{trader}{integer, the user id}

\item{trader}{integer, the user id}

\item{trader}{integer, the user id}

\item{date}{Date, the most recent time (included, i.e. not upto this date, but inclusive)}

\item{trader}{integer, the user id}

\item{date}{Date, the most recent time (included, i.e. not upto this date, but inclusive)}

\item{trader}{integer, the user id}

\item{date}{Date, the most recent time (included, i.e. not upto this date, but inclusive)}

\item{trader}{integer, the user id}

\item{date}{Date, the most recent time (included, i.e. not upto this date, but inclusive)}

\item{trader}{integer, the user id}

\item{date}{Date, the most recent time (included, i.e. not upto this date, but inclusive)}

\item{trader}{integer, the user id}

\item{date}{Date, the most recent time (included, i.e. not upto this date, but inclusive)}

\item{trader}{integer, the user id}

\item{date}{Date, the most recent time (included, i.e. not upto this date, but inclusive)}

\item{trader}{integer, the user id}

\item{trader}{integer, the user id}

\item{trader}{integer, the user id}

\item{trader}{integer, the user id}

\item{end}{Date, the most recent time (included, i.e. not upto this date, but inclusive)}
}
\value{
\code{keys} data.frame with generated keys.

\code{keys} data.frame with generated keys.

\code{keys} data.frame with generated keys.

\code{keys} data.frame with generated keys.

\code{keys} data.frame with generated keys.

\code{keys} data.frame with generated keys.

\code{keys} data.frame with generated keys.

\code{keys} data.frame with generated key.

\code{keys} data.frame with generated keys.

\code{keys} data.frame with generated keys.

\code{keys} data.frame with generated keys.

\code{keys} data.frame with generated keys.

\code{keys} data.frame with generated keys.

\code{keys} data.frame with generated keys.

\code{keys} data.frame with generated keys.
}
\description{
Generate data frame of start and end dates for a user.
Used to initialise ppmodel objects.

Generate three months lookback keys for user with
last day of previous month as last date

Generate key for last month for user with
last day of previous month as last date
and one month interval.

Generate three months lookback keys for user with
last day of previous month with regards to date as last date
and one month interval.

Generate four months lookback keys for user with
last day of previous month with regards to date as last date
and one month interval.

Generate twelve months lookback keys for user with
last day of previous month with regards to date as last date
and one month interval.

Generate three days lookback keys for user with one day interval.

Generate one month key for user.

Generate key for user with one year interval and one year lookback.

Generate keys for user with one year interval and three year lookback.

Generate key for whole year lookback from begining to end of date's year

Generate key for whole year lookback from begining to end of this year

Generate key for whole year lookback from begining to end of last year

Generate key for whole year lookback from begining to end of this year

Generate keys for whole including years start and end years
}

