% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\name{keyDiff}
\alias{keyDiff}
\title{Find set difference between stored keys and argument value
stores the result internaly}
\usage{
keyDiff(object, key_values)
}
\arguments{
\item{object}{object of class 'DataSet'.}

\item{key_values}{data.frame with columns corresponding to object key_columns and values of the keys to compare}
}
\value{
\code{object} object of class 'DataSet'.
}
\description{
Find set difference between stored keys and argument value
stores the result internaly
}
\examples{
keyDiff(object,key_values)
}

