% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keymap.r
\docType{class}
\name{KeyMap-class}
\alias{KeyMap-class}
\title{An S4 class to handle key generation}
\description{
Handles the relationship between queries for missing data rows
and thevalues of the URL fields that obtain that data
}
\section{Slots}{

\describe{
\item{\code{key_generator}}{"function"}

\item{\code{key_columns}}{"character"}

\item{\code{key_values}}{"data.frame"}

\item{\code{current_key}}{"data.frame"}

\item{\code{cntr}}{"numeric"}

\item{\code{fields}}{"character"}
}}

