% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_key_library.r
\name{last_month}
\alias{last_month}
\title{Generate key for last month for user}
\usage{
last_month(trader)
}
\arguments{
\item{trader}{integer, the user id}
}
\value{
\code{keys} data.frame with generated keys.
}
\description{
Generate key for last month for user with
last day of previous month as last date
and one month interval.
}

