% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_key_library.r
\name{last_year}
\alias{last_year}
\title{Generate keys for last whole year lookback from begining to end of last year}
\usage{
last_year(trader)
}
\arguments{
\item{trader}{integer, the user id}
}
\value{
\code{keys} data.frame with generated keys.
}
\description{
Generate key for whole year lookback from begining to end of last year
}

