% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectstore.r
\docType{class}
\name{ObjectQuery-class}
\alias{ObjectQuery-class}
\title{An S4 class implementing handling queries to objectstores derived from VirtualObjectStore.}
\description{
An S4 class implementing handling queries to objectstores derived from VirtualObjectStore.
}
\section{Slots}{

\describe{
\item{\code{fields}}{"character",}

\item{\code{values}}{"character",}

\item{\code{known_keys}}{"data.frame",}
}}

