% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sockets.r
\docType{methods}
\name{openConnection}
\alias{openConnection}
\alias{openConnection,ProcessSocket-method}
\title{open socket connection}
\usage{
openConnection(object)

\S4method{openConnection}{ProcessSocket}(object)
}
\arguments{
\item{object}{object of class 'ProcessSocket'.}
}
\value{
\code{object} object of class 'ProcessSocket'.
}
\description{
open socket connection
}
\section{Methods (by class)}{
\itemize{
\item \code{ProcessSocket}: open socket connection
}}

