% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_ppmodel_library.r
\docType{data}
\name{outcome_price_features}
\alias{outcome_price_features}
\title{List of outcome price features}
\format{An object of class \code{character} of length 5.}
\usage{
outcome_price_features
}
\description{
List of outcome price features
}
\keyword{datasets}

