% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_composite_datasets.r
\name{position_composite_factory}
\alias{position_composite_factory}
\title{Create and fill PositionComposite object}
\usage{
position_composite_factory(user, start, end, use_db = TRUE)
}
\arguments{
\item{user}{integer, trader ID.}

\item{start}{Date start date}

\item{end}{Date end date}

\item{use_db}{logical, pool data from Database or via Middleware}
}
\value{
\code{posn_comp_obj} object of class 'PositionComposite'.
}
\description{
Create and fill PositionComposite object
}

