% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access_sql_procedures.r
\docType{class}
\name{DataAccess.SQLProcedureCall.PositionHistory_SelectByTraderDate-class}
\alias{DataAccess.SQLProcedureCall.PositionHistory_SelectByTraderDate-class}
\title{DataAccess.SQLProcedureCall.PositionHistory_SelectByTraderDate" class}
\description{
Implements handling querries for raw trade data for given trader_id and date range
}
\details{
Inherits from "VirtualSQLProcedureCall"
}

