% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access_sql_procedures.r
\docType{class}
\name{DataAccess.SQLProcedureCall.PositionLevel_SelectFromHistoryByDate-class}
\alias{DataAccess.SQLProcedureCall.PositionLevel_SelectFromHistoryByDate-class}
\title{DataAccess.SQLProcedureCall.PositionLevel_SelectFromHistoryByDate" class}
\description{
Implements handling querries for trade position level data
}
\details{
Inherits from "VirtualSQLProcedureCall"
}

