% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access_sql_procedures.r
\docType{class}
\name{DataAccess.SQLProcedureCall.PositionService_SelectHistoryBetweenForStrategy-class}
\alias{DataAccess.SQLProcedureCall.PositionService_SelectHistoryBetweenForStrategy-class}
\title{DataAccess.SQLProcedureCall.PositionService_SelectHistoryBetweenForStrategy" class}
\description{
Implements handling querries for position history for given strategy
and date range
}
\details{
Inherits from "VirtualSQLProcedureCall"
}

