% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_ppmodel.r, R/models_ppmodel_library.r
\docType{class}
\name{PPModel-class}
\alias{AverageDownPsnGatherer-class}
\alias{PPModel-class}
\alias{PetNamesPsnGatherer-class}
\alias{PositionSizeBatchQuantiler-class}
\alias{PriceBySnapshotGatherer-class}
\alias{PrimaryPlacingPsnGatherer-class}
\alias{PsnAgeGatherer-class}
\alias{ResultsDayBatchGatherer-class}
\alias{ResultsDayExposureSnapShotBatchGatherer-class}
\alias{ResultsDayPsnGatherer-class}
\alias{ResultsDaySnapShotBatchGatherer-class}
\alias{SecondaryPlacingPsnGatherer-class}
\alias{SizeAdjustmentBatchGatherer-class}
\alias{StopTargetsGatherer-class}
\alias{TradeHistory-class}
\alias{TradeHistorySimple-class}
\alias{TradeHistorySimpleWithSummary-class}
\alias{TradeLevelGatherer-class}
\alias{TradeLevelSnapShot-class}
\title{An S4 class to represent generic ppmodel set.}
\description{
Pre Processing Model is an object that gathers and
processes raw data and aggregates it into one DataSet.
Depending on specific derived classes different result
can be obtained

PPModel derived class ResultsDayBatchGatherer

PPModel derived class ResultsDayPsnGatherer

PPModel derived class SizeAdjustmentBatchGatherer

PPModel derived class PositionSizeBatchQuantiler

PPModel derived class PrimaryPlacingPsnGatherer

PPModel derived class SecondaryPlacingPsnGatherer

PPModel derived class PrimaryPlacingPsnGatherer

PPModel derived class SecondaryPlacingPsnGatherer

PPModel derived class PetNamesPsnGatherer

PPModel class for PsnAgeGatherer

PPModel derived class TradeHistory

PPModel derived class TradeHistorySimple

PPModel derived class TradeHistorySimpleWithSummary

PPModel derived class PriceBySnapshotGatherer

PPModel derived class AverageDownPsnGatherer

PPModel derived class TradeLevelGatherer

PPModel derived class TradeLevelSnapShot

PPModel derived class StopTargetsGatherer

PPModel derived class ResultsDaySnapShotBatchGatherer

PPModel derived class ResultsDayExposureSnapShotBatchGatherer
}
\section{Slots}{

\describe{
\item{\code{queries}}{"WarehouseBatchQuery",}

\item{\code{warehouse}}{"NullableTradeWarehouse",}

\item{\code{ppdata}}{"DataSet",}

\item{\code{modeldata}}{"DataSet",}

\item{\code{preprocessor}}{"character",}

\item{\code{preproc_call}}{"nonstandardGenericFunction",}

\item{\code{features}}{"character",}

\item{\code{post_comp}}{"PPModelPostComputation",}

\item{\code{setup_comp}}{"PPModelSetupComputation",}

\item{\code{remove_wh}}{"logical",}

\item{\code{replace_features}}{"logical",}

\item{\code{warehouse_store_name}}{"character",}

\item{\code{index_by_date_column}}{"character",}

\item{\code{summary_comp}}{"PPModelSummaryComputation"}
}}

