% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_ppmodel_objectstore.r
\name{ppmodel_objectstore_factory}
\alias{ppmodel_objectstore_factory}
\title{Create PPModelObjectstore object}
\usage{
ppmodel_objectstore_factory(name)
}
\arguments{
\item{name}{"character", name of the objectstore}
}
\description{
Creates PPModel ObjectStore and loads data from
associated file if exists.
}

