% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_ppmodel_objectstore.r
\docType{class}
\name{PPModelObjectStore-class}
\alias{PPModelObjectStore-class}
\title{An S4 class implementing of PPModel Objectstore.}
\description{
Implements storage, queries, and update of PPModels
in an object and saving in related file.
}
\details{
Inherits from "VirtualObjectStore"
}
\section{Slots}{

\describe{
\item{\code{objectstore_q}}{"PPModelQuery"}

\item{\code{qry_store_nme}}{"character",}
}}

