% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_ppmodel_objectstore.r
\docType{class}
\name{PPModelQuery-class}
\alias{PPModelQuery-class}
\title{An S4 class handling queries to PPModelObjectstore.}
\description{
An S4 class handling queries to PPModelObjectstore.
}

