% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectstore.r
\name{.setObjectStoreKey}
\alias{.setObjectStoreKey}
\title{Set objectstore key}
\usage{
.setObjectStoreKey(object, objectstore_key)
}
\arguments{
\item{object}{object of class "VirtualObjectStore"}

\item{objectstore_key}{"data.frame" with key defining objectstore}
}
\description{
Private method to store object key
}

