% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_warehouse_objectstore.r
\name{.setObjectStoreKeyMap}
\alias{.setObjectStoreKeyMap}
\title{Set objectstore key_map object}
\usage{
.setObjectStoreKeyMap(object, key_map)
}
\arguments{
\item{object}{object of class "WarehouseObjectStore"}

\item{key_map}{object of class "KeyMap"}
}
\description{
Private method to store object key map object
}

