% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectstore.r
\name{.setObjectStoreQuery}
\alias{.setObjectStoreQuery}
\title{Set objectstore query object}
\usage{
.setObjectStoreQuery(object, objectstore_q)
}
\arguments{
\item{object}{object of class "VirtualObjectStore"}

\item{objectstore_q}{object of class "ObjectQuery"}
}
\description{
Private method to store objectquery object
}

