% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sockets.r
\docType{class}
\name{ProcessSocket-class}
\alias{ProcessSocket-class}
\title{An S4 class handling connections to R process socket.}
\description{
An S4 class handling connections to R process socket.
}
\section{Slots}{

\describe{
\item{\code{remote_host}}{"character"}

\item{\code{port}}{"numeric"}

\item{\code{server}}{"logical"}

\item{\code{socket}}{"NullableSocket"}

\item{\code{data}}{"character"}

\item{\code{read_limit}}{"numeric"}
}}

