% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_analysis_objectstore.r
\docType{methods}
\name{queryAnalysisStore}
\alias{queryAnalysisStore}
\alias{queryAnalysisStore,AnalysisObjectStore,data.frame-method}
\title{Query objectstore for given set of parameters}
\usage{
queryAnalysisStore(object, key)

\S4method{queryAnalysisStore}{AnalysisObjectStore,data.frame}(object, key)
}
\arguments{
\item{object}{object of class "DailyRiskModelObjectStore"}

\item{key}{"data.frame" wiht query keys}
}
\value{
\code{rval} object of class "AnalysisObjectStore" if query sucessfull,
otherwise NULL
}
\description{
Query objectstore for given set of parameters
}
\section{Methods (by class)}{
\itemize{
\item \code{object = AnalysisObjectStore,key = data.frame}: Query objectstore for given set of parameters
}}

