% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_trades_remote_objectstore.r
\docType{methods}
\name{queryClosestMatchFromTradeStore}
\alias{queryClosestMatchFromTradeStore}
\alias{queryClosestMatchFromTradeStore,TradeObjectStore,data.frame-method}
\title{Query store for Trade}
\usage{
queryClosestMatchFromTradeStore(object, key)

\S4method{queryClosestMatchFromTradeStore}{TradeObjectStore,data.frame}(object,
  key)
}
\arguments{
\item{object}{object of class "TradeObjectStore"}

\item{key}{"data.frame" with key related to query}
}
\value{
\code{rval} list of objects of class "Trade"
}
\description{
Querries Trade Objectstore for Trade stored under given key
Returns Trade if present NULL otherwise. Looks for closest match
to query
}
\section{Methods (by class)}{
\itemize{
\item \code{object = TradeObjectStore,key = data.frame}: Query store for Trade

Returns Trade if present NULL otherwise. Looks for closest match
to query
}}

