% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_daily_riskmodel_objectstore.r
\docType{methods}
\name{queryDailyRiskModelObjectStore}
\alias{queryDailyRiskModelObjectStore}
\alias{queryDailyRiskModelObjectStore,DailyRiskModelObjectStore-method}
\title{Query objectstore for given set of parameters}
\usage{
queryDailyRiskModelObjectStore(object, name, lookback, component)

\S4method{queryDailyRiskModelObjectStore}{DailyRiskModelObjectStore}(object,
  name, lookback, component)
}
\arguments{
\item{object}{object of class "DailyRiskModelObjectStore"}

\item{name}{"character" name of objectstore}

\item{lookback}{"integer" lookback value for model}

\item{component}{"character" name of the risk model component}
}
\value{
\code{rm} "DataSet" object with component if present otherwise NULL
}
\description{
Query objectstore for given set of parameters
}
\section{Methods (by class)}{
\itemize{
\item \code{DailyRiskModelObjectStore}: Query objectstore for given set of parameters
}}

