% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_ppmodel_objectstore.r
\docType{methods}
\name{queryPPModelStore}
\alias{queryPPModelStore}
\alias{queryPPModelStore,PPModelObjectStore-method}
\title{Query store for PPModel}
\usage{
queryPPModelStore(object, key)

\S4method{queryPPModelStore}{PPModelObjectStore}(object, key)
}
\arguments{
\item{object}{object of class "PPModelObjectStore"}

\item{key}{"data.frame" with key related to query}
}
\value{
\code{rval} object of class "PPModel"

\code{rval} object of class "PPModel"
}
\description{
Querries PPModel Objectstore for PPModel stored under given key
Returns PPModel if present NULL otherwise
}
\section{Methods (by class)}{
\itemize{
\item \code{PPModelObjectStore}: Query store for PPModel

Querries PPModel Objectstore for PPModel stored under given key
Returns PPModel if present NULL otherwise
}}

