% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_trades_remote_objectstore.r
\docType{methods}
\name{queryTradeStore}
\alias{queryTradeStore}
\alias{queryTradeStore,TradeObjectStore-method}
\title{Query store for Trade}
\usage{
queryTradeStore(object, key)

\S4method{queryTradeStore}{TradeObjectStore}(object, key)
}
\arguments{
\item{object}{object of class "TradeObjectStore"}

\item{key}{"data.frame" with key related to query}
}
\value{
\code{rval} object of class "Trade"

\code{rval} object of class "Trade"
}
\description{
Querries Trade Objectstore for Trade stored under given key
Returns Trade if present NULL otherwise
}
\section{Methods (by class)}{
\itemize{
\item \code{TradeObjectStore}: Query store for Trade

Querries Trade Objectstore for Trade stored under given key
Returns Trade if present NULL otherwise
}}

