% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_warehouse_objectstore.r
\name{queryWarehouseStore}
\alias{queryWarehouseStore}
\title{Query WarehouseObjectstore for given keys}
\usage{
queryWarehouseStore(object, trader_id, start, end)
}
\arguments{
\item{object}{object of class "WarehouseObjectStore"}

\item{trader_id}{"integer", trader ID}

\item{start}{"Date" start date of Warehouse}

\item{end}{"Date" end date of Warehouse}
}
\value{
\code{object} object object of class "WarehouseObjectStore"
}
\description{
Updates Warehouse if it doesn't contain data.
}

