% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_key_library.r
\name{range_years_lookback}
\alias{range_years_lookback}
\title{Generate keys for whole including years start and end years}
\usage{
range_years_lookback(trader, start, end)

range_years_lookback(trader, start, end)
}
\arguments{
\item{trader}{integer, the user id}

\item{start}{Date, the most distant time (included, i.e. not upto this date, but inclusive)}

\item{end}{Date, the most recent time (included, i.e. not upto this date, but inclusive)}

\item{trader}{integer, the user id}

\item{start}{Date, the most distant time (included, i.e. not upto this date, but inclusive)}

\item{end}{Date, the most recent time (included, i.e. not upto this date, but inclusive)}
}
\value{
\code{keys} data.frame with generated keys.

\code{keys} data.frame with generated keys.
}
\description{
Generate keys for whole including years start and end years

Generate keys for whole including months start and end months
}

