% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sockets.r
\name{readConnection}
\alias{readConnection}
\title{Read data from socket}
\usage{
readConnection(object)
}
\arguments{
\item{object}{object of class 'ProcessSocket'.}
}
\value{
\code{object} object of class 'ProcessSocket'.
}
\description{
Read data from socket
}

