% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_dataplex.r
\name{refresh}
\alias{refresh}
\title{Refresh Warehouse Cache}
\usage{
refresh(name)
}
\arguments{
\item{name}{warehouse objectstore name}
}
\description{
Creates new instance of WarehouseObjectstore and
stores it in cache
}

