% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_daily_riskmodel_objectstore.r
\docType{methods}
\name{reInitializeRiskModelComponents}
\alias{reInitializeRiskModelComponents}
\alias{reInitializeRiskModelComponents,DailyRiskModelObjectStore,character,integer-method}
\title{Reinitialize components to empty datasets}
\usage{
reInitializeRiskModelComponents(object, name, lookback = 150,
  components = NULL)


  \S4method{reInitializeRiskModelComponents}{DailyRiskModelObjectStore,character,integer}(object,
  name, lookback = 150, components = NULL)
}
\arguments{
\item{object}{object of class "DailyRiskModelObjectStore"}

\item{name}{"character" name of the model eg: "developed_europe_prototype"}

\item{lookback}{"integer" number of lookback horizon of the model, defaults to 150}

\item{components}{"character" name of the components of the risk model to be
re-initialized, possible values are:
c('ImpliedFactorReturns', 'ResidualReturns', 'Betas',
  'FactorCorrelation', 'FactorVariance', 'MarketStyle')}
}
\value{
\code{object} object of class "DailyRiskModelObjectStore"
}
\description{
Reinitialize components to empty datasets
}
\section{Methods (by class)}{
\itemize{
\item \code{object = DailyRiskModelObjectStore,name = character,lookback = integer}: Reinitialize components to empty datasets
}}

