% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_analysis_objectstore.r
\docType{class}
\name{RemoteAnalysisQuery-class}
\alias{RemoteAnalysisQuery-class}
\title{An S4 class handling queries to WarehouseObjectstore.}
\description{
An S4 class handling queries to WarehouseObjectstore.
}

