% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_objectstore.r
\docType{class}
\name{RemoteObjectQuery-class}
\alias{RemoteObjectQuery-class}
\title{An S4 class implementing handling queries to objectstores derived from VirtualRemoteObjectStore.}
\description{
Inherits from "ObjectQuery"
}

