% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_ppmodel_objectstore.r
\docType{class}
\name{RemotePPModelQuery-class}
\alias{RemotePPModelQuery-class}
\title{An S4 class handling queries to WarehouseObjectstore.}
\description{
An S4 class handling queries to WarehouseObjectstore.
}

