% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_trades_remote_objectstore.r
\docType{class}
\name{RemoteTradeQuery-class}
\alias{RemoteTradeQuery-class}
\title{An S4 class handling queries to WarehouseObjectstore.}
\description{
An S4 class handling queries to WarehouseObjectstore.
}

