% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\name{resetData}
\alias{resetData}
\title{Reset internal data storage to empty dataframe in the dataset}
\usage{
resetData(object, data)
}
\arguments{
\item{object}{object of class 'DataSet'.}

\item{data}{"data.frame" with new data}
}
\value{
\code{object} object of class 'DataSet'.
}
\description{
Reset internal data storage to empty dataframe in the dataset
}

