% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_daily_riskmodel_objectstore.r
\name{risk_model_objectstore_factory}
\alias{risk_model_objectstore_factory}
\title{Risk Model Objectstore factory}
\usage{
risk_model_objectstore_factory(name, lookback = 150)
}
\arguments{
\item{name}{"character" name of the model eg: "developed_europe_prototype"}

\item{lookback}{"integer" number of lookback horizon of the model, defaults to 150}
}
\value{
\code{rmstr} object of class "DailyRiskModelObjectstore"
}
\description{
Factory to create Risk Model Objectstore objects and load data if
stored in file
}

