% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\name{rowCoalesceOn}
\alias{rowCoalesceOn}
\title{aggregates underlying data on specified columns and using specified function}
\usage{
rowCoalesceOn(object, by = NULL, fn = first)
}
\arguments{
\item{object}{object of class 'DataSet'.}

\item{by}{character vector of column names on which to aggregate on, default are key columns}

\item{fn}{function used for aggregation, default is first()}
}
\value{
\code{object} object of class 'DataSet'
}
\description{
aggregates underlying data on specified columns and using specified function
}

