% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_ppmodel.r
\name{runPreProcessorModel}
\alias{runPreProcessorModel}
\title{run Pre Processor Model computation}
\usage{
runPreProcessorModel(object)
}
\arguments{
\item{object}{object of class "PPModel"}
}
\description{
run Pre Processor Model computation
}

