% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\name{setData}
\alias{setData}
\title{Sets new data in the DataSet}
\usage{
setData(object, data, override_full_index = FALSE)
}
\arguments{
\item{object}{object of class 'DataSet'.}

\item{data}{data.frame with data to be set.}

\item{override_full_index}{Should the index be reset? Defaults to FALSE}
}
\value{
\code{object} object of class 'DataSet'.
}
\description{
Sets new data in the DataSet
}

