% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_objectstore.r
\docType{methods}
\name{setSQLInsertObject,RemoteObjectQuery,BlobStorage.VirtualSQLProcedureCall-method}
\alias{setSQLInsertObject,RemoteObjectQuery,BlobStorage.VirtualSQLProcedureCall-method}
\title{Set sql_insert object in object slot}
\usage{

  \S4method{setSQLInsertObject}{RemoteObjectQuery,BlobStorage.VirtualSQLProcedureCall}(object,
  sql_insert)
}
\arguments{
\item{object}{object of class "RemoteObjectQuery"}

\item{sql_insert}{object of class "BlobStorage.VirtualSQLProcedureCall"}
}
\value{
\code{object} object of class "RemoteObjectQuery"
}
\description{
Private method to set sql_insert slot with "RemoteObjectQuery"
class object to be implemented in derived
}

