% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_objectstore.r
\docType{methods}
\name{setSQLQueryObject,RemoteObjectQuery,BlobStorage.VirtualSQLProcedureCall-method}
\alias{setSQLQueryObject,RemoteObjectQuery,BlobStorage.VirtualSQLProcedureCall-method}
\title{Set sql_query object in object slot}
\usage{

  \S4method{setSQLQueryObject}{RemoteObjectQuery,BlobStorage.VirtualSQLProcedureCall}(object,
  sql_query)
}
\arguments{
\item{object}{object of class "RemoteObjectQuery"}

\item{sql_query}{object of class "BlobStorage.VirtualSQLProcedureCall"}
}
\value{
\code{object} object of class "RemoteObjectQuery"
}
\description{
Private method to set sql_query slot with "RemoteObjectQuery"
class object to be implemented in derived
}

