% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_configs.r
\docType{data}
\name{switch_vars}
\alias{DATA}
\alias{ENV}
\alias{switch_vars}
\title{Global switch variables}
\format{An object of class \code{character} of length 1.}
\usage{
ENV

DATA
}
\description{
Global variable to control data environment

Global variable to control data sources
}
\keyword{datasets}

