% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_key_library.r
\name{three_monthly_lookback}
\alias{three_monthly_lookback}
\title{Generate keys for three months lookback for user}
\usage{
three_monthly_lookback(trader)
}
\arguments{
\item{trader}{integer, the user id}
}
\value{
\code{keys} data.frame with generated keys.
}
\description{
Generate three months lookback keys for user with
last day of previous month as last date
}

