% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_key_library.r
\name{three_year_lookback}
\alias{three_year_lookback}
\title{Generate keys for this whole year lookback from begining to end of today()'s year}
\usage{
three_year_lookback(trader)
}
\arguments{
\item{trader}{integer, the user id}
}
\value{
\code{keys} data.frame with generated keys.
}
\description{
Generate key for whole year lookback from begining to end of this year
}

