% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_trades_remote_objectstore.r
\name{trade_objectstore_factory}
\alias{trade_objectstore_factory}
\title{Create TradeObjectstore object}
\usage{
trade_objectstore_factory(key)
}
\arguments{
\item{key}{"data.frame", with keys for trade
required keys are: c("id", "instrument", "buysell", "strategy", "start", "end")}
}
\description{
Creates Trade ObjectStore and loads data from
associated file if exists.
}

