% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_trades_remote_objectstore.r
\docType{class}
\name{TradeObjectStore-class}
\alias{TradeObjectStore-class}
\title{An S4 class implementing of Trade Objectstore.}
\description{
Implements storage, queries, and update of Trades
in an object and saving in related file.
}
\details{
Inherits from "VirtualObjectStore"
}
\section{Slots}{

\describe{
\item{\code{objectstore_q}}{"TradeQuery"}

\item{\code{qry_store_nme}}{"character",}
}}

