% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_trade_factory.r
\docType{class}
\name{TradeWarehouse-class}
\alias{TradeWarehouse-class}
\title{An S4 class for storing trades}
\description{
Stores information about trades in given timespan for given trader.
}
\section{Slots}{

\describe{
\item{\code{trades}}{"environment",}

\item{\code{instruments}}{"numeric",}

\item{\code{features}}{"character",}

\item{\code{trader_id}}{"integer",}

\item{\code{positions}}{"PositionComposite",}

\item{\code{psn_summary}}{"DataSet",}

\item{\code{start_date}}{"Date",}

\item{\code{end_date}}{"Date",}

\item{\code{dly_data_pad}}{"numeric",}

\item{\code{map}}{"list",}

\item{\code{fctr_datstr}}{"character"}
}}

