% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_key_library.r
\name{two_year_monthly_lookback}
\alias{two_year_monthly_lookback}
\title{Function to generate date only keys over 2 year lookback
Generate keys for whole including years start and end years}
\usage{
two_year_monthly_lookback()
}
\value{
\code{keys} data.frame with generated keys.
}
\description{
Generate keys for whole including years start and end years
}

