% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_analysis_objectstore.r
\name{update_analysis_remote_storage}
\alias{update_analysis_remote_storage}
\title{Copy analysis blocks from local objectstores to remote store.}
\usage{
update_analysis_remote_storage()
}
\value{
\code{count} number of warehouses copied
}
\description{
copies all locally stored blocks to remote store and updates keys
fixes issues with wrong hashes
}

