% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_trades_remote_objectstore.r
\name{update_trade_remote_storage}
\alias{update_trade_remote_storage}
\title{Copy Trades from local objectstores to remote store.}
\usage{
update_trade_remote_storage()
}
\value{
\code{count} number of warehouses copied
}
\description{
copies all locally stored ppmodels to remote store and updates keys
}

