% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_warehouse_objectstore.r
\name{update_warehouse_remote_storage}
\alias{update_warehouse_remote_storage}
\title{Copy warehouses from local objectstores to remote store.}
\usage{
update_warehouse_remote_storage()
}
\value{
\code{count} number of warehouses copied
}
\description{
copies all locally stored warehouses to remote store and updates keys
}

