% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_analysis_objectstore.r
\docType{methods}
\name{updateAnalysisStore}
\alias{updateAnalysisStore}
\alias{updateAnalysisStore,AnalysisObjectStore,ANY,data.frame,logical-method}
\title{Update Analysis Store with new AnalysisObject}
\usage{
updateAnalysisStore(object, analysis_object, key, force = FALSE)


  \S4method{updateAnalysisStore}{AnalysisObjectStore,ANY,data.frame,logical}(object,
  analysis_object, key, force = FALSE)
}
\arguments{
\item{object}{object of class "AnalysisObjectStore"}

\item{analysis_object}{object of class "VirtualAnalysisBlock"}

\item{key}{"data.frame" with query keys}

\item{force}{"logical" should store be forced if object for given keys exists}
}
\value{
\code{object} object of class "AnalysisObjectStore"
}
\description{
Update Analysis Store with new AnalysisObject
}
\section{Methods (by class)}{
\itemize{
\item \code{object = AnalysisObjectStore,analysis_object = ANY,key = data.frame,force = logical}: Update Analysis Store with new AnalysisObject
}}

