% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\name{updateIndex}
\alias{updateIndex}
\title{Update index based on new incoming data}
\usage{
updateIndex(object, new_data)
}
\arguments{
\item{object}{object of class 'DataSet'.}

\item{new_data}{data.frame with new data to be added to dataset}
}
\value{
\code{object} object of class 'DataSet'.
}
\description{
Update index based on new incoming data
}

