% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_ppmodel_objectstore.r
\name{updatePPModelStore}
\alias{updatePPModelStore}
\title{Store PPmodel in Store}
\usage{
updatePPModelStore(object, ppmodel_object, key, force = FALSE)
}
\arguments{
\item{object}{object of class "PPModelObjectStore"}

\item{ppmodel_object}{object of class "PPModel"}

\item{key}{"data.frame" with key related to query}

\item{force}{"logical" force update of PPModel if it is already present}
}
\value{
\code{object} object of class "PPModelObjectStore"
}
\description{
Stores PPModel and reated Query in Store
}

