% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_trades_remote_objectstore.r
\docType{methods}
\name{updateTradeStore}
\alias{updateTradeStore}
\alias{updateTradeStore,TradeObjectStore-method}
\title{Store PPmodel in Store}
\usage{
updateTradeStore(object, ppmodel_object, key, force = FALSE)

\S4method{updateTradeStore}{TradeObjectStore}(object, ppmodel_object, key,
  force = FALSE)
}
\arguments{
\item{object}{object of class "TradeObjectStore"}

\item{ppmodel_object}{object of class "Trade"}

\item{key}{"data.frame" with key related to query}

\item{force}{"logical" force update of Trade if it is already present}
}
\value{
\code{object} object of class "TradeObjectStore"

\code{object} object of class "TradeObjectStore"
}
\description{
Stores Trade and reated Query in Store
}
\section{Methods (by class)}{
\itemize{
\item \code{TradeObjectStore}: Store PPmodel in Store

Stores Trade and reated Query in Store
}}

