% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_warehouse_objectstore.r
\name{updateWarehouseStore}
\alias{updateWarehouseStore}
\title{Update WarehouseObjectstore for given keys}
\usage{
updateWarehouseStore(object, keys)
}
\arguments{
\item{object}{object of class "WarehouseObjectStore"}

\item{keys}{"data.frame", keys to update for}
}
\value{
\code{object} object object of class "WarehouseObjectStore"
}
\description{
Updates Warehouse if it doesn't contain data.
}

