% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_warehouse_objectstore.r
\name{updateWarehouseStoreForKey}
\alias{updateWarehouseStoreForKey}
\title{Adds Warehouse to Warehouse Objectstore}
\usage{
updateWarehouseStoreForKey(object, new_warehouse)
}
\arguments{
\item{object}{object of class "WarehouseObjectStore"}

\item{new_warehouse}{object of class "CompositeWarehouse"}
}
\value{
\code{object} object object of class "WarehouseObjectStore"
}
\description{
Adds Warehouse to Warehouse Objectstore
}

