% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urldatareader.r
\docType{class}
\name{URLParser-class}
\alias{URLParser-class}
\title{class wraping XML Parser and ULR Reader}
\description{
implements methods used to parse between data retrieved from URL
in XML format and translates to R data.frame
}
\section{Slots}{

\describe{
\item{\code{reader_instance}}{"URLDataReader"}

\item{\code{class_instance}}{"XMLParser"}

\item{\code{urls}}{"character"}

\item{\code{objects}}{"environment"}

\item{\code{parser_type}}{"character"}

\item{\code{cstm_schm_tag}}{"character"}

\item{\code{cstm_schm_elem}}{"character"}

\item{\code{cstm_body_tag}}{"character"}
}}

