% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urldatareader.r
\docType{class}
\name{URLQuery-class}
\alias{URLQuery-class}
\title{An S4 class to represent generic URL query.}
\description{
An S4 class to represent generic URL query.
}
\section{Slots}{

\describe{
\item{\code{fields}}{"character"}

\item{\code{values}}{"character"}

\item{\code{root_url}}{"character"}

\item{\code{url}}{"character"}
}}

