% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.r
\docType{class}
\name{VirtualBaseQuery-class}
\alias{VirtualBaseQuery-class}
\title{An S4 class implementing basic query class and functions.}
\description{
An S4 class implementing basic query class and functions.
}

