% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectstore.r
\docType{class}
\name{VirtualObjectStore-class}
\alias{VirtualObjectStore-class}
\title{A VIRTUAL S4 class implementing basic functionalities of objectstore.}
\description{
Objectstore is an object that can store named datasets in separate environment
the datasets are stored and accessed using unique key values.
Objetstore also stores it's content to .Rds file and can be loaded later.
The API implements handling of reading and writing to the file.
}
\details{
Prototype is made only once and then copied with each call to 'new'
for objects along the same branch in the class hierachy.
Must ensure new environments created since they are passed
by reference.
}
\section{Slots}{

\describe{
\item{\code{stored}}{"environment"}

\item{\code{id}}{"character"}

\item{\code{data_path}}{"character"}

\item{\code{objectstore_q}}{"ObjectQuery"}
}}

