% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.r
\docType{class}
\name{VirtualQuery-class}
\alias{VirtualQuery-class}
\title{An S4 class implementing basic query accessors and functions.}
\description{
An S4 class implementing basic query accessors and functions.
}
\section{Slots}{

\describe{
\item{\code{fields}}{"character",}

\item{\code{values}}{"character",}
}}

