% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_dataplex.r
\name{warehouse_request}
\alias{warehouse_request}
\title{Query Warehouse for given keys}
\usage{
warehouse_request(name, trader, start, end)
}
\arguments{
\item{name}{"character" warehouse objectstore name}

\item{trader}{"integer" trader ID}

\item{start}{"Date" start date of warehouse}

\item{end}{"Date" end date of warehouse}
}
\value{
\code{warehouse} object of type TradeWarehouse if present NULL otherwise
}
\description{
Queries Warehouse for data for given keys.
Query goes via warehouse objectstore cache.
}

