% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_dataplex.r
\docType{data}
\name{warehouse_store_created}
\alias{warehouse_store_created}
\title{indicates if warehouse cache has been initialized}
\format{An object of class \code{logical} of length 1.}
\usage{
warehouse_store_created
}
\description{
indicates if warehouse cache has been initialized
}
\keyword{datasets}

