% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_warehouse_objectstore.r
\docType{class}
\name{WarehouseObjectStore-class}
\alias{WarehouseObjectStore-class}
\title{An S4 class for storing daily risk models.}
\description{
Inherits from "VirtualObjectStore"
}
\section{Slots}{

\describe{
\item{\code{key_map}}{"KeyMap"}

\item{\code{warehouse_q}}{"WarehouseQuery"}
}}

