% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_warehouse_objectstore.r
\docType{class}
\name{WarehouseQuery-class}
\alias{WarehouseQuery-class}
\title{An S4 class handling queries to WarehouseObjectstore.}
\description{
An S4 class handling queries to WarehouseObjectstore.
}

