% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sockets.r
\name{writeToConnection}
\alias{writeToConnection}
\title{Write data to socket}
\usage{
writeToConnection(object, data)
}
\arguments{
\item{object}{object of class 'ProcessSocket'.}

\item{data}{character, data to write to socket.}
}
\value{
\code{object} object of class 'ProcessSocket'.
}
\description{
Write data to socket
}

