% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urldatareader.r
\docType{class}
\name{XMLObject-class}
\alias{XMLObject-class}
\title{class wraping xml string and its parameters}
\description{
Default schema based on output of C# DataTable serialisation with CreateXML method
}
\section{Slots}{

\describe{
\item{\code{xml_schema}}{"ANY"}

\item{\code{xml_body}}{"ANY"}

\item{\code{schema_tag}}{"character"}

\item{\code{schema_element}}{"character"}

\item{\code{body_tag}}{"character"}
}}

