% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urldatareader.r
\docType{class}
\name{XMLParser-class}
\alias{XMLParser-class}
\title{class wraping XML Converter}
\description{
implements methods used to parse between R and XML
}
\section{Slots}{

\describe{
\item{\code{auto}}{"logical"}

\item{\code{converter}}{"XMLConverter"}
}}

