% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sockets.r
\name{closeConnection}
\alias{closeConnection}
\title{Close socket connection}
\usage{
closeConnection(object)
}
\arguments{
\item{object}{object of class 'ProcessSocket'.}
}
\value{
\code{object} object of class 'ProcessSocket'.
}
\description{
Close socket connection
}
\examples{
closeConnection(object)
}

