% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_ppmodel_objectstore.r
\docType{methods}
\name{commitPPModelStore}
\alias{commitPPModelStore}
\alias{commitPPModelStore,PPModelObjectStore-method}
\title{Commit store data to file}
\usage{
commitPPModelStore(object)

\S4method{commitPPModelStore}{PPModelObjectStore}(object)
}
\arguments{
\item{object}{object of class "PPModelObjectStore"}
}
\value{
\code{object} object of class "PPModelObjectStore"

\code{object} object of class "PPModelObjectStore"
}
\description{
Saves data stored in the object into file.
}
\section{Methods (by class)}{
\itemize{
\item \code{PPModelObjectStore}: Commit store data to file
}}

