% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\name{dataset_factory}
\alias{dataset_factory}
\title{Builds a dataset from a dataframe}
\usage{
dataset_factory(key_cols, data)
}
\arguments{
\item{key_cols}{character vector ofs column names to be used as identifying keys.}

\item{data}{data.frame to generate dataset.from}
}
\value{
\code{object} object of class 'DataSet'
}
\description{
Builds a dataset from a dataframe
}

