% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame_to_xml.r
\name{FrameToXML}
\alias{FrameToXML}
\title{Generates XML from data.frame}
\usage{
FrameToXML(df, name, tree_name = "Root")
}
\arguments{
\item{df}{data.frame to convert}

\item{tree_name}{name of the xml tree, default is 'Root'}

\item{character}{name}
}
\value{
\code{xml} character xml string
}
\description{
Generates XML from data.frame
}
\examples{
rowCoalesceOn(object,by=NULL,fn=first)
}

