% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\docType{methods}
\name{getData}
\alias{getData}
\alias{getData,DataSet-method}
\title{Return data stored in the DataSet}
\usage{
getData(object)

\S4method{getData}{DataSet}(object)
}
\arguments{
\item{object}{object of class 'DataSet'.}
}
\value{
\code{data} data.frame with raw data.
}
\description{
Return data stored in the DataSet
}
\section{Methods (by class)}{
\itemize{
\item \code{DataSet}: Sets new data in the DataSet
}}

