% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_daily_riskmodel_objectstore.r
\docType{methods}
\name{getMostRecentRiskModelBetasDate}
\alias{getMostRecentRiskModelBetasDate}
\alias{getMostRecentRiskModelBetasDate,DailyRiskModelObjectStore-method}
\title{Get most recent betas date stored in object}
\usage{
getMostRecentRiskModelBetasDate(object, name, lookback = 150)

\S4method{getMostRecentRiskModelBetasDate}{DailyRiskModelObjectStore}(object,
  name, lookback = 150)
}
\arguments{
\item{object}{object of class "DailyRiskModelObjectStore"}

\item{name}{"character" name of the model eg: "developed_europe_prototype"}

\item{lookback}{"integer" number of lookback horizon of the model}
}
\value{
\code{date} "Date" date of the most recent betas and factor variances.
Returns "NULL" if nothing is stored
or only components that are not model specific are stored.
}
\description{
Get most recent betas date stored in object
}
\section{Methods (by class)}{
\itemize{
\item \code{DailyRiskModelObjectStore}: Get most recent betas date stored in object
}}

