% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_trade_factory.r
\docType{methods}
\name{getRawPositionData}
\alias{getRawPositionData}
\alias{getRawPositionData,TradeWarehouse-method}
\title{Get position data from Warehouse}
\usage{
getRawPositionData(object)

\S4method{getRawPositionData}{TradeWarehouse}(object)
}
\arguments{
\item{object}{object of class "TradeWarehouse"}
}
\value{
\code{positions} object of class "DataSet" with position data

\code{positions} object of class "DataSet" with position data
}
\description{
Returns all position data stored in Warehouse
}
\section{Methods (by class)}{
\itemize{
\item \code{TradeWarehouse}: Get position data from Warehouse

Returns all position data stored in Warehouse
}}

