% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_trade_factory.r
\name{getRawPositionSummary}
\alias{getRawPositionSummary}
\title{Get position summary data from Warehouse}
\usage{
getRawPositionSummary(object)
}
\arguments{
\item{object}{object of class "TradeWarehouse"}
}
\value{
\code{psn_summary} object of class "DataSet" with position summary data
}
\description{
Returns all position summary data stored in Warehouse
}

