% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_warehouse_objectstore.r
\name{getWarehouseFromStore}
\alias{getWarehouseFromStore}
\title{Get warehouse from WarehouseObjectstore for given keys}
\usage{
getWarehouseFromStore(object, trader_id, start, end)
}
\arguments{
\item{object}{object of class "WarehouseObjectStore"}

\item{trader_id}{"integer", trader ID}

\item{start}{"Date" start date of Warehouse}

\item{end}{"Date" end date of Warehouse}
}
\value{
\code{wh} object of class "CompositeWarehouse" if present otherwise NULL
}
\description{
Retrieves from WarehouseObjectstore warehouse matching
keys passed as arguments.
}

