% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_trade.r
\docType{methods}
\name{initialize,VirtualTrade-method}
\alias{initialize,VirtualTrade-method}
\title{Initialize method for "VirtualTrade" class}
\usage{
\S4method{initialize}{VirtualTrade}(.Object, order_id, leg_start, leg_end,
  trader, instrument, strategy, long, buysell, value_usd, consolidation, status)
}
\arguments{
\item{.Object}{object of class "VirtualTrade"}

\item{order_id}{"integer" order ID of the trade}

\item{leg_start}{"Date" date of start of trade leg}

\item{leg_end}{"NullableDate" date of end of trade leg}

\item{trader}{"character" integer id if the trader as character}

\item{instrument}{"integer" ID of the traded instrument}

\item{strategy}{"character" name of the traded strategy}

\item{long}{"logical" is Trade Long or Short}

\item{buysell}{"character" buy or sell c("Buy", "Sell")}

\item{value_usd}{"numeric" value of trade in USD}

\item{consolidation}{"data.frame" with data related to trade leg.
should contain columns : c('TradeDate','ValueUSD','Strategy', 'OrderID')}

\item{status, }{object of class "VirtualTrade"}
}
\value{
\code{.Object} object of class "VirtualTrade"
}
\description{
Initialize method for "VirtualTrade" class
}

