% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_trades_remote_objectstore.r
\name{key_from_trade_objectstore_name}
\alias{key_from_trade_objectstore_name}
\title{helper function to generate key from objectstore name}
\usage{
key_from_trade_objectstore_name(name)
}
\arguments{
\item{name}{"character" name of the entry in objectstore}
}
\value{
\code{key} "data.frame" with columns 'id', 'instrument', 'buysell', 'strategy', 'leg_start', 'leg_end', 'status'
}
\description{
helper function to generate key from objectstore name
}

