% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_key_library.r
\name{key_generator}
\alias{key_generator}
\title{Generate data frame of start and end dates for a user.}
\usage{
key_generator(trader, end, interval, lookback, calendar = TRUE,
  lookback_unit = "months", cols = c("id", "start", "end"))
}
\arguments{
\item{trader}{integer, the user id}

\item{end}{Date, the most recent time (included, i.e. not upto this date, but inclusive)}

\item{interval}{integer, the date decrement}

\item{lookback}{integer, the number of decrements including the end date}

\item{calendar}{logical, if calendar is true, it is assumed that end is the final day of a
calendar period (eg. month,week) and the date intervals are adjusted to span the calendar interval.}

\item{lookback_unit}{character, name of time unit as recognized by lubridate:
c("days", "weeks", "months", "years"), default is "months"}

\item{cols}{character, name output columns, default is c('id','start','end')}
}
\value{
\code{keys} data.frame with generated keys.
}
\description{
Generate data frame of start and end dates for a user.
Used to initialise ppmodel objects.
}

