% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sockets.r
\name{openConnection}
\alias{openConnection}
\title{open socket connection}
\usage{
openConnection(object)
}
\arguments{
\item{object}{object of class 'ProcessSocket'.}
}
\value{
\code{object} object of class 'ProcessSocket'.
}
\description{
open socket connection
}

