% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_composite_datasets.r
\name{position_composite_factory}
\alias{position_composite_factory}
\title{Create and fill PositionComposite object}
\usage{
position_composite_factory(user, start, end,
  source = .__DEFAULT_POSITION_HISTORY_DATA_SOURCE__.)
}
\arguments{
\item{user}{integer, trader ID.}

\item{start}{Date start date}

\item{end}{Date end date}

\item{source}{"character", where the data should be pulled from}
}
\value{
\code{posn_comp_obj} object of class 'PositionComposite'.
}
\description{
Create and fill PositionComposite object
}

