% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_ppmodel.r, R/models_ppmodel_library.r
\docType{class}
\name{PPModel-class}
\alias{AverageDownPsnGatherer}
\alias{PPModel-class}
\alias{PetNamesPsnGatherer}
\alias{PositionSizeBatchQuantiler}
\alias{PriceBySnapshotGatherer}
\alias{PrimaryPlacingPsnGatherer}
\alias{PsnAgeGatherer}
\alias{ResultsDayBatchGatherer}
\alias{ResultsDayExposureSnapShotBatchGatherer}
\alias{ResultsDayPsnGatherer}
\alias{ResultsDaySnapShotBatchGatherer}
\alias{SecondaryPlacingPsnGatherer}
\alias{SizeAdjustmentBatchGatherer}
\alias{StopTargetsGatherer}
\alias{TradeHistory}
\alias{TradeHistorySimple}
\alias{TradeHistorySimpleWithSummary}
\alias{TradeLevelGatherer}
\alias{TradeLevelSnapShot}
\title{An S4 class to represent generic ppmodel set.}
\description{
Pre Processing Model is an object that gathers and
processes raw data and aggregates it into one DataSet.
Depending on specific derived classes different result
can be obtained
}
\section{Slots}{

\describe{
\item{\code{queries}}{"WarehouseBatchQuery",}

\item{\code{warehouse}}{"NullableTradeWarehouse",}

\item{\code{ppdata}}{"DataSet",}

\item{\code{modeldata}}{"DataSet",}

\item{\code{preprocessor}}{"character",}

\item{\code{preproc_call}}{"nonstandardGenericFunction",}

\item{\code{features}}{"character",}

\item{\code{post_comp}}{"PPModelPostComputation",}

\item{\code{setup_comp}}{"PPModelSetupComputation",}

\item{\code{remove_wh}}{"logical",}

\item{\code{replace_features}}{"logical",}

\item{\code{warehouse_store_name}}{"character",}

\item{\code{index_by_date_column}}{"character",}

\item{\code{summary_comp}}{"PPModelSummaryComputation"}
}}

