% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access_sql_procedures.r
\docType{class}
\name{DataAccess.SQLProcedureCall.Query_HistoricalTrades_WithInstrumentIDAndOrderID-class}
\alias{DataAccess.SQLProcedureCall.InstrumentHistoryRequired_QueryPriceHistoryFromTQA-class}
\alias{DataAccess.SQLProcedureCall.Query_HistoricalTrades_WithInstrumentIDAndOrderID-class}
\title{DataAccess.SQLProcedureCall.Query_HistoricalTrades_WithInstrumentIDAndOrderID class}
\description{
Implements handling querries for raw trade data for given trader_id and date range

Implements handling querries for raw trade data for given trader_id and date range
}
\details{
Inherits from "VirtualSQLProcedureCall"

Inherits from "VirtualSQLProcedureCall"
}

