% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_ppmodel_objectstore.r
\name{queryPPModelStore}
\alias{queryPPModelStore}
\title{Query store for PPModel}
\usage{
queryPPModelStore(object, key)
}
\arguments{
\item{object}{object of class "PPModelObjectStore"}

\item{key}{"data.frame" with key related to query}
}
\value{
\code{rval} object of class "PPModel"
}
\description{
Querries PPModel Objectstore for PPModel stored under given key
Returns PPModel if present NULL otherwise
}

