% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_warehouse_objectstore.r
\docType{methods}
\name{queryWarehouseStore}
\alias{queryWarehouseStore}
\alias{queryWarehouseStore,WarehouseObjectStore-method}
\title{Query WarehouseObjectstore for given keys}
\usage{
queryWarehouseStore(object, trader_id, start, end)

\S4method{queryWarehouseStore}{WarehouseObjectStore}(object, trader_id, start,
  end)
}
\arguments{
\item{object}{object of class "WarehouseObjectStore"}

\item{trader_id}{"integer", trader ID}

\item{start}{"Date" start date of Warehouse}

\item{end}{"Date" end date of Warehouse}
}
\value{
\code{object} object object of class "WarehouseObjectStore"

\code{object} object object of class "WarehouseObjectStore"
}
\description{
Updates Warehouse if it doesn't contain data.
}
\section{Methods (by class)}{
\itemize{
\item \code{WarehouseObjectStore}: Query WarehouseObjectstore for given keys

Updates Warehouse if it doesn't contain data.
}}

