% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sockets.r
\docType{methods}
\name{readConnection}
\alias{readConnection}
\alias{readConnection,ProcessSocket-method}
\title{Read data from socket}
\usage{
readConnection(object)

\S4method{readConnection}{ProcessSocket}(object)
}
\arguments{
\item{object}{object of class 'ProcessSocket'.}
}
\value{
\code{object} object of class 'ProcessSocket'.

\code{object} object of class 'ProcessSocket'.
}
\description{
Read data from socket
}
\section{Methods (by class)}{
\itemize{
\item \code{ProcessSocket}: Read data from socket
}}

