% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_ppmodel.r
\docType{methods}
\name{runPreProcessorModel}
\alias{runPreProcessorModel}
\alias{runPreProcessorModel,PPModel-method}
\title{run Pre Processor Model computation}
\usage{
runPreProcessorModel(object)

\S4method{runPreProcessorModel}{PPModel}(object)
}
\arguments{
\item{object}{object of class "PPModel"}
}
\value{
\code{object} object of class "PPModel"

\code{object} object of class "PPModel"
}
\description{
run Pre Processor Model computation
}
\section{Methods (by class)}{
\itemize{
\item \code{PPModel}: run Pre Processor Model computation
}}

