% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_trade.r
\docType{class}
\name{Trade-class}
\alias{Trade-class}
\title{Concrete S4 class for storing trade info.}
\description{
Stores information about trade leg together with
all necessary "features" that can be attached
}

