% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sockets.r
\docType{methods}
\name{writeToConnection}
\alias{writeToConnection}
\alias{writeToConnection,ProcessSocket-method}
\title{Write data to socket}
\usage{
writeToConnection(object, data)

\S4method{writeToConnection}{ProcessSocket}(object, data)
}
\arguments{
\item{object}{object of class 'ProcessSocket'.}

\item{data}{character, data to write to socket.}
}
\value{
\code{object} object of class 'ProcessSocket'.

\code{object} object of class 'ProcessSocket'.
}
\description{
Write data to socket
}
\section{Methods (by class)}{
\itemize{
\item \code{ProcessSocket}: Write data to socket
}}

